it was working well but data coming from matlab was wrong :(
