+++ src/const_config_GF64_N64.hpp
@@ -14,7 +14,13 @@ using gf_type = uint16_t;
 using value_type = float;
 #define max_gf 65535
 
-// this relialbity is for SNR=-7.5dB
-constexpr uint16_t reliab_seq[GF] = {61, 60, 59, 63, 55, 62, 47, 31, 57, 58, 53, 51, 45, 54, 46, 43, 39, 29, 27, 30, 23, 49, 56, 52, 50, 41, 42, 44, 38, 15, 37, 28, 35, 26, 25, 22, 21, 48, 19, 40, 14, 36, 13, 34, 24, 33, 11, 20, 18, 7, 17, 12, 10, 32, 9, 6, 16, 5, 3, 8, 4, 2, 1, 0};
-const value_type chan[GF * N] = {1.18837513887877,1.68579597396990,1.82857029729434,1.57324427761549,0.611496007516679,0.619823319538108,1.18695380754264,1.18488853836523,1.19216302681775,1.11569512697692,0.922767106146492,1.51007533103167,1.69745670990956,0.945823460876231,1.84076407707452,1.63009463079767,1.92875875431464,1.24245476561668,1.26868032092623,1.00552570311447,1.54892848152875,1.26398518197932,1.64299462109159,1.27958121341877,0.697042564901864,1.27333080212735,2.44475251804909,1.51436869793769,0.342547730518337,0.960911052042725,2.24887198681205,0.783555478027709,1.69334622106922,2.43991047813285,1.69460969661749,0.683283511141497,1.70498113723561,1.09916189689694,0,1.21172102332192,0.788486272883520,1.57434332734251,1.81408130501282,1.62202417346371,0.876314385094922,1.40929229903420,1.34529724997496,0.308046756430624,0.789708475140684,1.54530187571751,1.31638271377908,0.179685811489091,1.30688964672670,0.460501524158841,1.64347554265431,1.86533211542767,0.394713323516434,1.57797939098060,1.26170330092912,0.559509411265705,0.0607954188289646,0.895114276318000,1.34720440701450,1.20274123536427,1.76916775835111,0.671978523164179,0.965089235896956,1.03733224980927,0.162090748487794,0.921956975669387,1.47032141588188,1.15063544500719,1.18508898341093,1.32622239903752,0.788913305998455,1.72649765528360,0.124542456866063,0.446377350337118,1.75063850722784,1.57642356169274,1.19118640735428,0.623329686784476,1.11897141331993,1.15839242481366,0.788359509299885,0.444080488809219,1.86658424281501,0.971675447397163,1.22928133631526,1.20611192499565,0,0.968297609638147,1.54223012612547,1.78762425811993,1.10443247278927,1.61902753063688,0.762668390160766,0.813527682967881,1.56493366153833,0.869171338927940,1.17709515312345,0.945306321106893,0.781631167499906,1.25946952291845,1.10876891503934,1.04563713024125,1.46550889618966,0.677889476971782,0.747596497568447,0.949850268064201,0.562888076772766,1.01163632322044,0.659538966384515,1.91768216859586,0.710522252347960,1.47943205161581,0.793149272990402,0.744508150598717,1.20172666750543,1.68961153362580,1.09164049621578,0.508023499096844,0.802479717120276,1.36520459992098,1.50110867184141,0.574967769334093,0.875251722899177,0.738205832840315,1.70457136053606,2.40715765862136,2.71601555869498,2.12820904746457,2.29521997121057,2.51301422462728,1.57349032421604,1.89880629248045,3.72728890684263,1.42030130027336,2.10941763894323,2.00614640309770,1.94101599871718,2.44288124410297,2.50608439408726,2.36441715449875,2.36455638594757,1.59747551332777,2.58322966797424,2.13292643096222,2.61223778360398,1.92923384748919,0,1.74297983262161,2.45713130405271,2.44452303828574,3.07354405266297,1.28165397109124,2.40747653498892,2.73671153103616,1.85567179651189,2.82733584061102,2.18609963739395,2.58821145044239,2.11906990031199,2.22016274037790,1.96445666498885,2.60844536700960,2.79803042005852,2.14111674492056,2.59196292810492,2.68458888681119,2.07809183953290,1.82817730905590,1.97617267615711,2.06798458285019,2.10497223030030,2.58478651579446,2.16711454030266,2.15466598735085,2.03743919089624,1.91615507284913,2.75693389367519,2.58885296067038,2.93324915892884,1.36855371430842,2.56462605228054,2.91284484341847,3.09212796976836,2.11522562040217,1.10640472228571,2.29593391258143,2.00097567917878,2.44492019132830,1.46143396716059,1.57530567183845,2.60939895720013,1.81199712778301,2.26706515701527,1.61207572581162,1.51754889507149,1.81621281428276,3.15511643710242,2.07048281272155,1.65685448017740,1.39351920228091,1.22895351189290,1.59538406277684,2.14747605704647,0,2.49341450406808,2.41677551975864,2.22497543520573,1.60899939535383,2.05883424928446,1.78458714131393,2.50038418049006,1.75254791669246,2.49595287219112,1.60453447066070,1.62060286397286,1.97937146899177,2.15062569392534,1.38844109342226,2.28999061093232,1.51938111018828,2.04230638551829,1.84476942577586,2.41594203215089,2.50062421930711,2.28615873765660,2.20435505067490,1.31625554260487,2.21076479412031,1.84522568466571,2.35456135523788,1.87773763671422,1.03460115874197,1.10745531718079,2.57817917293883,2.42798110190020,1.86517059458706,3.00280850201916,1.24237952126076,2.15024596417195,1.69190169660691,2.06106434917288,2.76546129412954,1.27446282529836,1.73499955196861,1.47359242209147,1.89700674359650,2.41025616117499,1.62997302706403,2.23024898748096,2.83928914305327,2.43219567298338,1.62711516307759,0.621865521265895,0.240745253048143,1.92039200976186,1.05947524177778,0.697268432950473,0.652830217004128,0.175021512050501,0.657152205848494,0.777958937975569,1.14393314748333,0.698164274086601,0.657588080787997,0.541124774106623,0.690673873855819,0.553263833997808,0.345472278297196,0.743600888782806,1.15407393330951,0.288562349734941,1.32448441279757,0.0163110153215733,1.42049107059233,0.896696122491017,0.965841697902612,1.64550114412355,1.68203183818759,0.708611058739165,1.76377265335107,0.402714215941611,0.979917859210814,1.35009244229452,0.911157013122352,0.972474485854325,0.689962114635262,0,1.31165867924711,0.626693892196094,1.92052130983622,0.946184363544485,0.568380530167672,0.277461141325329,1.02169631041005,0.516592312652545,0.986262860724305,0.667068582441747,1.43624624288399,0.978543387754794,0.656551673069444,0.326862460935814,0.830855100557944,1.38763165140936,1.42843688989092,1.71512446322056,0.884369370623124,0.243621728855584,0.653013273079284,0.673305740426613,1.22454450616407,1.18181122972495,1.66567065875859,0.653465050915727,0.871575480520340,1.27306276940901,0.754922907134455,1.67982909832826,1.33727548488341,1.63217003381906,1.38613864059135,1.39038864570915,2.67128557553753,1.92027199480711,1.41042815210452,0.817903761417054,1.34214650788986,2.22210640685087,1.86383254003800,1.24113710830579,1.10023995580507,1.08732793958249,1.29059820089210,2.48186661045321,1.64708882334274,3.00194774574071,0.840194740265188,1.70240396598474,1.74995489600366,1.26685466913547,1.37288136458665,1.79255908404229,1.18210154380755,1.01998275701053,1.64672673305776,1.40910605087258,1.16805955692839,0.971864295658833,1.79226246125401,1.10206489417088,1.35982412937311,1.65235911536048,1.49423733690943,1.02244720431549,0.810402710625875,1.97652755946842,0.477120712327631,2.34351479756058,1.94459573796947,1.95635713211252,0.909284555323131,1.08999924546221,1.68709359004751,1.72567543632890,0,1.62491581282188,2.65696136255421,0.663580919469050,1.49960827451914,0.793264589849564,1.14273438410016,1.80398213635492,1.48364527010173,0.712599688079747,0.527496065079616,0.431278547011628,1.79027429690845,2.15912876839899,1.80524355358055,1.23911007710462,1.21891876459486,1.69304345876069,1.25148462432521,1.83838351376509,1.46105279203268,0.418786605860865,1.75094081472844,1.42797252505929,0.346501855125416,0.648562672860537,1.68405283593070,1.09476160618030,1.34580340673156,1.19187760023190,1.58271410788571,0.329965109537258,1.45883619154304,2.46860199958907,0.362662832628104,1.24824774339946,1.02238303395503,0.818870474936030,1.96781864496871,0.822265158112986,0.376853267149029,1.86794461589991,0.970297014126409,1.23361062400360,0.532801820690954,1.35417174197803,1.36671507573701,0.226843822198408,1.56435782151545,1.50032727781787,0.890476086958238,1.83262497831386,1.38146041781871,1.21279578030281,0.908178801825703,0.691920234111650,0.949604947733036,1.93867173108464,1.48098533981184,1.59750241523370,1.08550876275413,0.590331716184350,0.782393031410722,1.29921496921998,1.86851140234220,1.44483682763648,0.444056669754266,2.09289991076776,1.81062257093363,0.922876031001084,1.26586215443255,0.713702070361552,1.00761636397549,1.67040464866355,1.56632500000600,1.03817841826982,0,1.45633194665439,1.04021949277719,1.96898915722403,1.60250906730608,2.19248216444233,1.65142931897748,3.24569728844703,1.92958090514814,2.60949628603010,1.08904564039477,2.12751179431004,2.00906834975999,1.78109165422744,2.42060932370083,2.10804968552995,1.48758742698852,2.41173036610803,1.91684566086974,1.21490386692882,1.94789953469430,1.04695980446996,2.41312365545112,2.01578060690136,2.34196672015192,2.27907148580937,1.14021031367972,2.44771606074524,2.69498141015641,1.96835093502991,2.88569875386705,1.96870526399073,1.74611871284994,2.48215613929458,2.06326209027696,1.69521089380629,2.21300645776211,2.92080564050585,2.38918995755161,2.25704608771350,1.95368617281760,2.24305040886700,1.20344489254938,1.93851137057711,2.56698515362337,2.21364727487420,2.62768003922000,1.96014090348757,0,2.89794105802348,2.05306055522260,1.50753790998414,1.73565268140548,1.96576812177774,2.24230403575231,2.17054132676035,2.07409557509289,1.79627735302304,1.67640202236655,2.36059081933524,2.66852414586839,1.90093990486359,1.92141226584861,1.31205550831825,2.30635944672799,1.92216358384158,2.39514589866496,1.14016709370836,2.23290071892803,1.66938645430499,2.40349712431036,0,1.78296176972461,2.32079056706871,2.51260854028798,1.52593030731667,2.71876489204954,1.44710802595928,2.79674847949202,2.47532298730731,2.77667941947564,1.84163829391437,2.78707069956322,1.52419189895608,2.98316712779961,2.48215303820879,2.34567021677669,2.34100786735787,1.89250414007827,2.24850134171424,0.813046213875998,1.63815670969705,1.98343301476786,1.89659930849486,1.80912904474854,2.91389967142596,2.39967456810131,1.74352277734940,2.31817703906030,1.40766512134743,2.43122380521920,2.28941919032395,2.06913629678712,1.92571821739769,1.95986370541083,3.09017058213871,2.78031734352310,2.01400429123767,2.61732221068179,1.34273768218278,2.92684676411870,1.79133340998168,1.67802218352851,1.93632541306852,2.62125494518657,2.23044940407412,2.76382388419102,2.32676944435071,2.18138067734314,2.76847444310932,3.54856833440662,2.52753250133504,2.57743721575162,1.84576848466054,1.67825388093471,2.69822543369633,1.66333824396125,2.57430711830014,2.46989422656282,2.74049805324185,2.24475888457933,2.52694014159849,1.54254068444941,1.54896175531897,2.10369404451304,1.35009437880800,1.73948301282911,0.930561530291752,1.15143522156795,1.18615099049871,2.20050088959804,0.986784943166922,0.685622781176242,1.22992148386511,0.750441488798813,0.812410549960192,1.75475737175070,0.723046585748452,2.30660320505904,2.05935484392855,2.15984126877487,1.32232837494040,0.853788658865791,0.914872474686379,1.53930960362443,1.64087463360861,1.47853237179649,0.839229971446922,0,0.850301737785022,1.82670862288559,1.10986761511433,0.905032175045998,2.63905184053524,1.29475924050809,1.41259447923495,1.22355638987519,0.625011122526758,3.10581947648056,2.27157289137211,2.07044458786826,1.61900128565242,1.72687509769360,1.41810969966477,1.64798611816292,1.37719126850607,1.06808290540756,2.07017432755060,1.66076219131370,1.64303017943911,1.07744358117137,0.793056917440171,1.05625849288664,1.71266419921324,1.07170349533624,1.19787284011206,1.80446260390002,1.31238913634536,1.69450226077667,1.77813277986609,1.98400020672560,1.22571954693013,1.30026136334471,1.50647425497696,0.251474492363815,0.879392211127429,2.19140789140222,1.82288020111932,2.50029462926678,2.38258391388963,1.89000393382081,2.49750028415046,1.88073787088305,2.37924938481511,2.27466980141111,2.39429432499328,1.56441110220851,2.14442054625588,1.73000299392195,2.52403298791190,2.35754325985660,1.90582908763154,1.48097085668888,2.54940103248228,2.81920333913726,2.27571885806565,2.92729807036048,2.27787635383326,3.51381641962015,3.41586656233967,1.56747824597631,2.05441537181154,1.26089518680373,2.58832115298574,3.09784636979659,2.42478192347306,1.39677494602650,2.13838323322310,1.50978215105051,2.27571848555144,1.88891952795853,1.93020115963536,2.14227456432270,2.30699447981464,2.58719499458718,1.34316817520417,0,2.02030642770411,2.06357506451419,3.59422476078616,1.84570142993568,2.79300641914142,1.85317656633567,1.53205812901803,2.24776308357055,3.05307605942789,2.14757678120882,2.17740111687240,1.91166793963328,2.31806459782564,2.67032759751689,1.37283915463340,2.41497526112281,1.98050192214519,2.69483639762796,1.46000644129589,2.72584913975921,2.50302582526746,1.72264629526813,1.22677128981397,2.96557711600745,0.908121235266393,1.57136888942397,0.469624193255732,1.17519525512315,1.22619716173658,0.712592972455126,0.346929146603086,1.30172139646210,1.37109902403577,1.05490073052782,0.829886498639535,0.549280309254966,1.07170206457192,2.04412162527989,1.65523260925643,0.240390362575671,0.850554286116017,0.857872964909224,1.64590971890861,1.39826784349801,1.12676087168262,0.954510591347143,1.06703077008107,1.20057110854428,1.40281758198464,1.97882321732440,1.95386334001546,0.884935464648638,1.75511869112013,0.861410960313612,0,1.71357500289062,1.30479825652681,1.18517526972279,1.49528520604204,0.893611070615823,1.15235455370893,1.91966825128226,1.25519187896081,0.943457808255589,1.38022883727343,1.25644096277965,0.611016478702372,1.43259464248409,0.776508868709879,0.434690857689201,1.77005284193756,1.93799868284726,0.180985893870645,1.50636110137849,1.45971486061856,0.820895433977932,0.838993217778638,1.59206385909593,0.854535665938838,1.77762235855604,1.36973857429492,1.33269272006514,1.43097440053506,1.96939114132864,0.830920856985128,0.575323807175663,1.36280236183621,1.54572815644881,3.60064772654570,1.89561047377957,1.68861297143005,2.83426482586177,2.10756766388842,3.06217224534775,0,3.50678029151754,1.98077185539638,2.02879550179542,2.15717932157871,1.81808715437189,1.48030486701348,3.29448579782051,2.70014031701677,3.39667781494278,2.55564490188865,3.24918716461209,3.65284340309281,2.44226467496602,1.25164971111378,1.86872659541409,2.76753409305298,3.36512858566265,2.38924064079877,2.21561979934480,2.22174100931692,2.30228335222973,2.63891848189673,2.89211076197591,2.08166979925656,2.93000389715985,2.22383093683029,3.70439022396762,1.74208142835249,2.51562409731029,2.11683394628366,3.01796445112717,4.26770142779410,2.66557090890093,2.31127301659265,3.12797673624258,3.42896708361036,2.14093623414978,2.78973081371975,2.65984961395472,3.73669546140088,2.85225635387804,2.47427916231019,1.95568694217551,2.50676804099416,2.23242169130459,2.25569035966021,2.89869926628913,2.65547569196370,2.24786047707980,2.07285171238912,3.27350967225999,2.20785188271601,3.23983209285463,2.83212564221797,2.35831393897803,2.16919735923669,2.57205238637032,2.13416729498867,1.15661755617323,1.72707478142543,2.48874604417537,1.56443406143162,1.45319792513552,2.32866257389689,1.32523086302016,2.31455821241440,2.06383144906303,2.08171600437883,1.75326416513028,1.75263485223122,2.97076900562773,1.93465871711129,2.61305745288246,1.97775297463107,2.38968178721223,1.14363964670934,2.00104979409751,2.05006963216555,2.78936216777977,2.37798027047519,0.808248855544679,1.84053890752547,2.64501830232430,1.87385613297463,1.99528554494260,1.95511923831904,1.86630124550314,1.63937002412367,1.80482038508643,1.81293117419450,1.60748952949544,2.52993509590306,1.48284964370945,2.14606627316116,2.18824033654795,1.02935634801806,1.30914406461928,1.96249687499119,2.35840419984731,1.93904571492654,0,1.72170345700644,1.72922311689427,1.67886383227565,1.32909541585324,1.84827731209826,3.05569924109500,1.57622419276742,1.39698889811393,2.06043654641341,2.27785951100617,2.57215734726134,1.12352163005135,2.17550495770081,1.93492865891069,1.92841213624634,2.23607585189072,1.92376836468906,1.56508944322167,1.63482151243628,2.12229212341320,1.78678957202487,0,1.68729425058635,1.32049887930762,1.48889692457192,2.13033665356496,1.87334860556101,0.738467481870720,3.17486511283850,1.92324368091086,1.67478793840083,1.52176830449834,2.08646662535954,2.23576936677660,1.54284188285797,1.30682533945327,1.81034722210442,2.53234331614733,2.43983609440940,1.17609296830844,0.870037820756365,1.62404493619339,1.45955451031560,1.47593123855081,2.20639142070685,2.48000267590546,2.12734648215151,1.68191751297263,2.33906289492134,1.84863288486272,1.41902780631585,1.21177099218603,1.27906208334965,2.06407130906918,1.53309662253254,1.97650023430173,1.93674571933307,2.16156468814705,1.67515158577534,1.15421779898216,2.24558104120096,2.09018844531669,2.17898892145522,1.28717585048206,2.02004818244183,1.76909840101037,1.78037897462442,1.72335295057630,1.95307849698786,2.55948347806069,1.66989483946657,1.47274117693902,1.20479126959452,1.80644277182128,1.77620177573096,0.940164926044871,1.56327518725354,2.11755758411362,2.53421960850277,1.80083044373073,1.57876131965693,1.68568413157788,1.68028628484313,1.20258318472063,2.43507612611943,2.98260661622213,2.84447836420476,3.09234506463932,3.17675457509321,3.21496703837744,3.26346187574324,2.60051568343384,3.27688562865028,2.76702582058685,2.86959358971674,2.59053606269280,3.45393704660883,2.96445657444913,2.72318640814612,2.08591508770551,3.33087700155298,2.70583038307752,2.19800560548508,3.29557890114978,2.61589616361083,3.70429547353344,3.26371185546568,2.15491744044513,2.44485256950834,2.95007427458872,3.11852946532288,2.66277668070482,3.08754460146943,2.81736374677571,2.94144569586265,2.69111418741034,3.47036590101779,1.85846956003230,2.56984882766634,2.70261317981395,2.88176013848668,2.86311764575457,2.80462567452951,0,2.78055612952802,4.09278823527556,2.35764278394976,2.71433499795572,2.93416531286439,2.89585216164658,3.41398430159600,3.23794296505979,2.90924110878725,2.69475732014371,3.08070194823574,3.15223318384293,2.92824569008424,3.39070667235679,2.76022233476688,2.89109928847594,3.48895400920517,2.55249909894560,2.25886056430601,3.60198103056234,3.10844666389596,2.70425703703542,3.06031470875299,2.98094963734579,1.77256915421511,1.15519209251537,2.12453420678138,1.86124910928383,0,1.71924028558754,1.85512624376835,1.80463583041833,1.35041865182065,1.82885905712694,1.42000897476814,2.09183707412195,0.442866906872517,0.687183018548747,2.06661674725538,2.50538750430899,2.30313943930072,1.76030263047913,1.69735187513532,2.05266413283505,1.70400628671449,1.97107864224737,2.58663200373402,1.83533841049536,2.08408943197944,1.53539234988187,0.907154352459830,2.18618239219792,1.96717632602395,0.876409859370884,2.39202578344511,1.75494358404557,2.18096756949007,0.810795786602306,2.40755490186744,3.12271734659768,1.02000067311680,1.47288714630522,2.40983893693975,1.92401851485133,1.61058119444158,1.73030470114475,2.36822172487822,1.12280989854631,1.99796949825547,1.35135397882584,2.09922009573602,1.34193708669873,1.34298290980103,2.18113406282751,2.63814233255795,2.92556207356622,2.11150091852743,1.62763408921495,0.789415532942742,1.85805291761715,1.24021792278164,2.07765157280338,2.17822078577744,1.63701331530053,2.07479999633463,1.44158912013814,1.27245517168827,1.94353453307777,1.82262580386000,1.34367083576520,2.09105929059243,1.84028998685935,0,1.24647634082887,1.99365857298057,1.05454440125724,1.81876731337298,2.21419769085022,2.13051335412801,1.58930022588946,1.72598986436893,0.751202041380271,1.79998819727628,0.824840686780976,2.19016607706566,1.37833122874191,1.95126635761131,0.717323454172452,1.65175120788031,1.99472417011055,1.70634654306640,2.40177475431617,1.58347158719700,2.83586845165561,0.487639992045466,1.32676699827924,1.20631828010052,0.843210454229131,0.932303112488561,2.05460464845430,1.65159200383361,2.02830857895947,0.688964085363395,1.04122248402922,2.05822786207950,1.34402700923884,1.97117054919935,2.08516849635348,0.601816130017705,0.702578518111265,1.41184258281059,1.61518549776875,2.43402751634694,1.45671615536437,1.26319838128666,1.81604146041869,0.976904692859461,1.01591485909825,0.989847612044444,1.08220184708834,2.28314680691771,0.805710567232246,0.928558682364212,2.00714174218530,2.23077931335088,0.775427196094590,1.66773090730754,2.27346016616857,1.00007531181822,1.12557544001696,1.43253163607237,1.65249014142892,1.17224410795148,0.445522588172057,0.466560369280404,1.49614689599989,1.58450584788514,0.776725601762384,1.64692086799523,1.79784688710511,0.866144401929013,1.82941955163987,0.821579007978705,0.685638761647534,0.834959448375744,1.22574761658183,1.14785833920599,0.568660504944148,0.688923308584709,0.274770529878008,0.880287276543117,0.286262758696712,0.952587565488736,0.742711030579379,1.36168129096303,0.580744651092944,1.52810669647794,1.02256634844821,1.01218953668736,1.84850121344167,0.799847613222491,2.03640842001910,1.19122490477173,1.84126852645969,0,1.25704454611797,0.672951073147742,1.20102897898799,1.66659767291331,1.65014159965735,1.09665845089250,1.21975966863148,1.61983362876206,0.358442723794882,1.20603734752098,1.28801329501749,1.26974250808724,1.78027998269738,1.79256455126076,0.415609252660332,2.03472977860186,1.25871062464361,1.08604027411274,1.32878213289036,0.698934912949612,0.988709773589323,1.46730970592187,1.76569936900132,1.21667262229473,1.33165798030577,1.72898145978289,1.00581451953476,0.547456686742337,0.994716441988488,1.58256117314241,0.917123198989001,1.61065963009532,2.56104537424528,1.26660261107271,0.877253437960267,2.00488224237515,0.962184437022492,1.32553849628198,2.00106033566365,1.93650231376352,1.68262464315214,1.29499769913115,0,2.23670575478192,1.42806665463467,1.57723017342553,1.29081952783533,1.55876347864709,2.09619928580749,1.88166461304981,2.70933397950504,1.86440002725100,1.77268467474669,1.47879604528173,2.40130046680740,1.25561705919160,1.84399025047358,1.39374673470205,1.62321658943515,2.36441182265054,1.84537403215951,1.11572754303226,2.13136050407431,2.45128997656654,2.56695336395260,1.84498600358897,1.86326428896400,1.36058336682357,2.41456742557632,1.42847979856773,1.82276825559150,1.46943753892566,1.33441016883753,2.81017090611773,1.66988898416322,1.71710381080877,1.71474456799045,1.50857609456271,1.67458177326273,1.76573761362578,1.89323164980268,1.59527336945029,1.26843953689452,1.89504797722851,2.85196815548970,1.54322343909672,1.35284923770447,0.762894015545676,2.65267582358098,1.57175067935046,1.20598807161404,1.11261043257559,1.30538490900363,1.39187527749224,2.03767119240638,1.46381649821622,1.68685308459764,1.74232644686200,1.76609974553315,1.62992564710168,2.48953587723169,1.64848443437329,2.32665168545476,1.66161655803257,2.38459402134551,1.67482093749711,2.07627864758674,1.11378583967274,2.49499360186939,1.33911124853228,1.64809915742463,1.16664884976250,1.73542306326159,1.96069469784132,1.47276374067755,1.90468220704940,1.55636294194565,1.85161141158397,1.32962628826601,1.07488917530090,1.42477249043812,1.68305472444607,1.76986400762313,1.27336857057643,2.04066661732165,2.32313073309064,2.55275284724056,1.26314034004723,1.36780489558760,1.10244536471630,1.80951064361937,2.25502248061460,2.35484689744125,1.78495480725214,1.35311991100801,1.11248123008463,1.50022096782421,2.09224108876756,1.79340224601004,1.24488033937069,2.00853860005759,1.47698752007214,2.10531496572632,1.76311301645712,2.62265117558189,1.29151258309019,1.50696425156651,0,1.88628103961952,1.21832116543298,2.20998631938699,0.707400312226343,2.22532245621039,2.11183978338913,2.08331978852701,1.69156963399084,2.04570064452633,2.55429183049459,2.20371934567226,1.74924090525553,2.72608569701507,1.16599826751540,1.90772574753264,1.75478875115735,2.25489793775120,2.65549495557565,0.548818642013412,0.736000478768909,1.64901513647252,1.76064998243223,2.23221394605107,1.89934281298483,1.50033569323391,1.52062663982191,1.57477419110571,2.51548004600120,1.46270633058009,2.57929378976741,0.845055751035515,1.44668829690101,1.45676378068499,1.21014982491458,0,2.09333901315385,3.40846720262166,2.08082405758951,1.38264935678533,1.37241836367403,2.71496205431899,1.77830028480600,2.56960861357712,2.27639499198160,1.50942598350836,1.41896299061097,1.83883714116496,2.26961849529012,2.81130853109461,1.77064181004401,2.42502045881885,2.65495214144029,2.57699464491114,1.20157604896632,1.54384117170625,1.66853724149403,1.95425359193997,1.20374134267424,1.99276006201777,1.01536322922204,1.73746298184914,1.82432019665998,1.88159516911758,1.29999162202690,0.612822517715013,1.95481162847940,2.81074160629415,1.15059518101802,1.00253545361722,0.984615262459412,1.32118756190218,1.87497188463156,2.19231818577795,1.28979254583630,0.922473608553002,2.00394277459506,2.14950990214108,2.01439102670892,1.99650259324679,1.03978940466954,2.31694230822648,2.39592077007870,2.18969588994617,2.18149193428567,2.35639915198393,2.06488600538751,1.92156906853883,2.27543869281135,1.83883973869219,1.66339483579506,2.60488244704704,2.06155634705366,1.63621785849782,1.44440695311080,1.56405740018144,1.72188069819077,2.32156791495569,2.04433987155636,1.58068117282305,2.23980708834381,1.16274732673749,1.70146591574686,1.78191422909428,1.96290223510775,1.63054128659103,2.58945285218789,1.81046874312346,0.950156903855926,1.93144668413175,1.27739092791470,2.12503735777805,1.97264226194557,1.91475260729091,1.45487915669399,2.53976519472138,2.36297717409033,1.59024504198997,0.430651735511246,2.29079641911785,1.88802656731985,1.56653212399278,0,1.88801495655175,2.36796868208152,1.49692658120770,2.01860137413020,1.83458110313272,0.751321860595842,1.58523746073317,1.99093317216564,2.55684503303746,2.16887755458654,1.88531784280215,1.84083031673472,1.05265822886625,1.77354499934464,2.35595142971209,2.39811449044317,1.60165425001541,1.44671530917403,2.00184500805490,1.87650661036349,1.84742606142627,0.883157830356497,1.92460441700931,0,2.02096300135767,1.02699244433564,1.49010500174179,0.801946196252849,1.44133369463879,2.08780121615528,1.92954985098856,1.91676925208276,0.968221507235258,2.14786655508353,0.896021641294811,2.22125589584653,1.92066043575570,1.75141560150284,2.29592578332192,1.59387992960459,1.56286798131849,1.47926240036849,1.34669948881726,1.44489770036922,2.29274674113551,1.83427923019071,0.970929372719404,1.10660875657736,1.18362034414878,1.23321640759622,2.01046441016064,0.993464765765512,2.28602236384200,1.73590419149973,2.22959170321651,2.19658367419939,1.75471782606583,1.66951209938311,2.21155037420546,1.51877534270033,1.76254793993287,1.32198409715645,1.97242239794331,1.68298357805813,1.59077845462148,1.27996236147766,1.10529552654221,2.31968232287122,2.15778857896912,2.27096160160226,2.17874777003375,2.15168205925527,1.98258156309364,2.28912726977662,1.93444061869817,1.87297321154306,1.79749865914290,1.65741934296394,1.67105402975349,1.08467631615637,0.998938003026502,1.82908079424236,1.63321915543355,2.69187987510135,2.25552216689328,1.73720282680756,0.914107695089512,2.76957520892721,2.45862796805593,2.26851149164526,2.04155962045480,2.02954825230110,1.95408777989392,2.25292992541129,2.21698665970536,1.63307761663214,1.98168843754673,3.55205317967559,1.95323806876145,1.90917729161389,2.43376961120307,2.79909239365691,2.00060030971023,2.04733691125222,0,2.41485810126580,2.06548852258120,3.03672935027356,1.09314772875781,1.85266548691981,0.885863019790243,1.94040736396803,2.43771481936432,2.48130611043072,1.98347877698765,1.74951784042156,2.20889846730808,1.65359115424222,1.72940949244009,2.70929828716036,2.04875797746345,1.94213193595672,2.00535450767117,2.51846310741545,1.52091879409216,2.60720547395240,1.34802150683198,1.91507201868385,1.95423019534904,2.85116764512003,1.75999958767436,2.21974810829959,1.30192654186798,2.48918945969749,1.38949267401697,1.47555991682415,2.04940904869956,2.36372867731697,2.27859838894524,3.26313042659503,1.65919859742410,2.00874770760416,1.56549650167838,1.79171330302542,1.30440954952233,2.04529609136458,1.39922431505288,1.11990443005845,1.17993005997249,1.52467471162896,1.23986060694651,1.09703702031566,2.20605378410721,0.938799750441956,1.38632872285859,1.05533066221401,0.892650852442289,1.86246742049828,2.04156234334917,1.99258225413973,1.42001320745845,1.37653714322364,1.69109508442397,1.75298181042484,1.58101726452578,2.44184238545790,2.53991735881923,2.10302615923873,1.52847942123087,1.38575559757300,1.36855108340907,1.33775487329151,1.35374525293358,1.32837081707200,2.74116448103133,0.382655598824940,2.19103846415233,2.08149362199503,2.34724521379858,1.76843682110525,2.27683478403094,1.76919287762231,1.77495603364547,1.22885327254936,0,2.03462253458607,0.659982859007880,1.98965167481807,1.42841809766245,1.66062868118813,2.29598189277497,1.52794186961985,1.74404099605873,1.98575004346522,1.74543841376257,1.84651063403194,1.15011231856537,1.05901698792307,1.57919175952579,1.61514982858103,1.12804197165259,1.32754644487730,1.94851355115360,1.79541812168708,1.79815908471830,1.74676007219942,1.97286526823850,2.14336734762842,1.36231959846735,1.57419042086023,2.12687981524323,2.11484011999696,1.91947341832332,1.23390934938204,1.77929352359727,2.55943031161432,1.35185956041423,2.41639345883374,1.37819347526356,1.37372877927475,2.24366922499239,1.27644783172668,1.54560128322090,0.572474076296623,1.40555322443051,1.50958976593090,1.71941169816436,1.77352012910448,1.41841822258770,1.79648146552948,1.23177249470230,2.29434199848218,2.04923978854475,1.62231795784537,1.39394468072118,0.411183871666370,2.35873013426705,2.44603533700507,1.41267472857045,0.946776397216729,1.48203417260258,1.82242632937467,2.26514555889046,1.84399282636093,0.749864977062405,2.55652415180769,2.25265502115606,1.55812396232716,0.888190228009656,1.24973503138585,1.27587894879606,2.08407821519462,1.30162896593822,1.74187332891825,0.847195787780185,2.40237660819677,2.11078616862830,1.08471700729301,1.73524740667135,2.19022790616426,2.30916620584330,1.60757732471794,0.977639060103231,0,1.23544216119492,1.34908723494224,1.11697154672887,0.477684331429474,1.80006721109775,1.42192218272962,1.94342629992609,1.68807614450762,1.93247618259085,1.30286933338326,0.603239697216328,1.09258577020525,0.550218768286236,1.35517765263922,1.77842450097070,1.50034347978645,1.17192046421464,1.03967948874626,0.962729694674284,1.77169519505420,1.67816832569839,0.886134513035873,0.690729903627287,1.39182701801104,1.37574645745382,0.551848695568220,0.529151299758928,0.989504455123733,1.98876563408928,1.14863982284983,1.93614481725415,0.814441685540797,0.456651689443182,1.92247823115519,1.84601829022057,0.757816413487886,0.746476941263461,1.35814333447528,1.60579816739851,2.21035432715655,0.896314885698069,0.724089742842890,1.34691896109234,1.47896671676772,1.32989802283132,0.902018015684250,1.76850532059863,0.772144705150316,1.91246253819426,0.870304254877380,0,0.987369174100471,2.13866111233800,1.29949358875955,1.29789254105904,0.581181988648888,1.05514233756456,1.08618615170180,0.498430856850494,1.68447877509839,1.32367084277011,1.16538565524563,2.16968609365326,0.360643221442570,0.446475345322247,0.474701191208838,1.66878211342237,0.944294627841819,1.87838398350863,0.573454092226958,1.11192017445022,1.72494444486106,1.33759348452296,0.884872852496872,0.476401174479263,0.333716487957735,0.943138245732770,1.36583072565061,1.39738635021340,1.52322865971375,1.55257489318885,1.35291563431454,0.696551921860226,1.55421203723458,1.43175026981052,1.69907504557478,0.969876689793226,2.07682996625838,1.92730456843831,0.939353974952993,1.32128715449738,2.20459724995439,2.16889439080595,1.63941737594714,0.282727008360414,0.463815349091478,1.14505757563376,1.75142246886277,0.960881432162450,0.742383564183425,0.515700395227526,2.07719778020807,1.22438175281111,0.760346068821335,1.02547367433222,1.06471674616443,0.407426387049847,1.26725869489951,0.996098511574377,0.880964217292473,0,1.17130408576653,1.15515499727793,1.83404224008816,1.04531622739702,1.84835486259644,1.27251958786993,0.761418706166545,0.346563678131245,1.36798997934990,1.82689153701017,1.52626555795623,0.719768393563099,2.00124773973072,1.30147505383210,0.470503180827054,1.53670509834291,1.04937975637667,1.81864196163486,0.517082426095536,0.233629808967514,1.27015754642700,0.964788778512543,1.50375559677501,2.17566518188460,1.71578982621269,0.499929386566723,2.24094946350680,1.94583392039389,1.63437744209252,1.81735171640471,2.09803092902172,2.10787111788208,1.81556731104316,2.25551125934521,2.86701021503854,2.58583273545306,2.60786996486529,1.11794413339405,2.42517746021439,2.91477611049903,2.50353862599138,2.14075808013210,1.84488283666723,2.39548542115642,2.91035178154752,2.01703688745173,2.56307932912732,1.58379483614132,3.80540786688786,2.21372222313600,2.79402199955017,2.08361582894296,3.00318838595029,2.60586320181610,2.46568338331482,2.10154576527462,1.78777077281601,3.32949004774158,2.52716690968339,2.41729418120175,2.93025471055449,2.28973530309647,2.58389119773846,2.33042523841032,2.90742138601330,3.68835169397060,2.62894234004955,1.68429977604563,1.77559038336861,2.36139287224274,3.02788168507353,1.56581161372739,2.55949741218132,1.58513271271871,2.79267850104721,3.13648031405479,2.58602504468969,1.38686398149488,2.43835755610006,3.47280032212625,2.80985393175628,1.81711832557372,0,2.24184544408552,2.85929199687327,2.43786254383057,1.73838933544929,1.32635290553049,1.93933762360782,2.59120251290481,2.27035230873978,1.87325915822922,2.01081115558082,1.82353852353471,1.50087772860045,1.03297063041148,2.19287679035789,2.09604554034251,1.13467192856018,2.51887597182764,1.93780806952332,3.03602166015569,2.36828985004491,1.93186692580737,1.97672125948391,1.85082078338993,1.68803439713641,2.18291869459402,1.80186253939183,1.70525197613275,2.26318634383196,1.65062622153165,2.37842813131811,2.35057593819367,1.28109818078474,0.647144880195725,2.84637414013953,1.57618950932045,1.41059577529159,2.18886367794352,1.85064789935871,2.03943884434563,1.75016167109585,1.18933672999126,1.68478413608218,1.90825371072373,1.13942014399486,2.53745546289469,2.09156132328138,2.08061230490782,1.44816565302782,1.36391682859765,2.48186912336788,1.73122385967818,0,2.09242230525914,1.48600819725446,1.94706645187755,1.44448696339883,0.979556787975960,1.94450968075607,1.85507098271915,2.19944673688716,2.00837878938894,2.39245245159041,2.29250424490575,1.90611130198484,1.29499874340984,2.11926637957622,1.55348732136155,1.76391994933382,2.62085159588586,1.48704396799562,1.19724103631395,2.06357374833796,0.538879706646164,1.67097852535747,0,1.85555766456194,0.684312425863441,1.72805274022564,0.728626001821510,1.71803227940772,2.00883732864589,1.22714759393167,0.887088364462256,1.07655491585959,0.620575660141096,1.79200792579850,0.362003920889817,1.37160164594710,1.22368650558959,1.50116174214850,2.12860189593695,1.50955862049828,1.24394515977736,1.87635818061741,1.46052058426129,1.80088138953068,1.91914053202305,0.265543604554463,1.07151982402486,0.455525891174657,0.867347609365575,1.37764661526305,0.367897290452450,0.960833884857875,0.960871247342152,0.982672944321779,1.64060475876715,0.805188162728525,1.27629440803803,1.27356476496211,1.58749857749268,1.30535796277393,1.24610007972953,0.963529007940687,1.23178677706017,1.38995534652231,1.00767558168183,1.79163700501270,2.17635616306210,1.73281928602051,1.64065713872457,1.64404099904999,1.48021486506504,0.558489271464787,1.19928343364875,0.894966022224468,0.479638281518998,1.68718520433577,1.58114739652861,1.80031944476496,1.59676820771447,0.554574406230638,1.42254066931869,0.956292296587075,2.20633903524919,1.16306522319312,1.81543890323688,1.99778805998599,1.33610965535508,1.54263472216720,2.72038966664917,1.10536467608389,1.43850028562255,1.22908500888230,2.03113699798641,2.43595545420720,1.94970383001156,2.21191584725818,1.83657326451449,1.73118133933887,1.22020891256161,1.85160706340885,1.94065028255400,1.74997073504179,1.90759899176795,2.36978044197335,2.43254470127391,1.10493249329652,1.83419226206729,1.94181785860343,1.10809565859585,2.02311286831580,1.43849352730412,2.53991955478028,1.88047115459040,1.68761220834655,1.21367469149572,2.53680126757697,1.39794598692146,1.42481375471882,1.51826234796033,1.88416784549492,1.38393814954072,2.38664253390768,2.45179323002713,1.66014869472208,1.10034595722584,1.11664517989146,1.36016822622947,1.35661247452900,0.818912550920266,0,1.80684131426473,1.17578644754653,1.56489814324374,1.72816042632961,1.50430229278714,1.57398167115081,1.51368139846547,2.02478737727523,1.67716462343391,0.291518162457916,1.12075978042220,1.32362516638540,1.86058736078379,1.98281724747747,1.93514468803452,0.847288044489616,3.16018955197707,1.67390287984814,1.32663671682935,1.27894085540781,1.83565047066684,0.652789906402790,2.18640495639981,1.59056396693273,1.82374673538905,1.97512071794462,1.03761567576861,1.32136700291952,1.75046027019062,1.73472071761096,1.34805066248928,0.978465323187354,0.540408552275588,1.30851067320291,1.03397089786338,1.33049006981112,1.80982050073068,0.532000285070525,1.58417401575907,2.27845138101663,1.51766329231058,1.61430682263932,0.479379306660605,0.919113039532395,0.602959332266608,0.547439638226026,1.59935579039373,1.47727719485709,0.710863210079300,0.407459886916606,1.06332767579865,1.82003092193383,1.67790235467805,1.48918542266405,0.530074783294086,0.808385325896195,1.18576481695469,0.729638932259031,0.989565686667271,1.82882402263150,0.698970527538112,1.12243584042932,1.39492481382219,0.511252773485682,1.71291199498843,0.334528665292855,0.888035290843758,1.95967519275307,2.35293170791430,1.36195641142317,0.349230501587287,1.08757461127086,1.77725640462728,1.88653849657094,1.35353741574971,0,0.653251838495729,2.30376937534657,1.67268342737168,1.50882812575740,1.13100210922575,2.56763299304871,0.974415619862060,2.10706355486554,1.73217152513106,2.47147397667372,0,2.40968989493344,1.99714113622061,1.55593441668168,1.27898008931902,0.558566906210533,2.17232777887643,2.07486274137647,2.78939973461429,1.98422707592004,1.43461759403128,1.21303587241941,1.13306619608570,2.52865828169348,1.65056068674563,0.958155817758193,1.00993653133304,2.42366268507497,2.03360537708748,0.864444495766889,2.39244957861197,1.54252566115848,1.59086763133385,0.831568645605348,1.98623171884936,1.30245484358573,0.887669716057777,1.97313494376710,2.13187650609185,2.57363492375682,0.811668427215169,1.86968592515117,1.29500626376925,1.31552884661034,2.47536782480420,1.73006750365925,1.08625794204755,2.03429761936796,1.75000199557220,1.67944230893795,1.62385055282262,1.22179487416734,1.35700302958564,2.48502553882423,2.63301853337968,1.73904304614133,1.61468211255397,2.55631462730530,1.62827164455757,1.40510566433995,2.24574285321610,2.17136149526015,1.78275061682212,1.77172022288914,1.29384163763424,2.90975248995578,1.09458975991766,2.11467609154442,2.09803695308982,0.606854327969219,1.84309202883582,1.38025435986338,1.25676195530157,0.902725802586358,1.41630196941890,1.09750070215750,2.07924612087399,1.52879025158848,1.72575713823820,1.02994920542522,0.842385327943597,2.25123547084597,0.984695797389409,0.809965535103142,0.984662180381783,1.38188679246564,2.55886891944459,1.77789396222651,0.920879551879907,1.16064519259005,1.98870144652676,1.86273349766189,1.85781488689924,1.12383476529301,0.984729064267317,1.02531519427176,1.63818542091047,2.02392500892883,0.495385970392906,1.94650894880920,2.24254699165744,2.46545354072093,1.49487643556249,0,1.04954544048977,1.97809641252477,2.17737170609307,1.46210566744609,0.751287705801019,1.98433824901985,1.23797581274082,2.07217735088977,1.09764973088918,1.39452433540046,0.968451391842280,1.32142946150889,2.22598094287756,1.26238030368941,1.58292667453335,1.23260585809648,1.24510909367655,1.68883593007454,1.71664193349663,1.36598421822629,1.46076452807544,2.20146920598758,1.64719568880710,1.45207132323615,1.19527536340314,1.29437132320319,1.56904985133325,0.606480403058783,0.267710914333182,1.21608393651223,0.113307061921953,0,0.992241478309910,0.629915393788080,0.204713746985913,0.859621148880331,2.04701458076676,0.940708713647582,1.74300358392903,0.627965878187971,1.10835977090513,1.09941395338637,0.926942068027747,0.865798133866457,1.38015174195321,0.679863072097809,1.23005448115683,0.139472218921772,1.48559033553512,1.09628232755798,1.63446789262469,0.321112707235893,0.588725056191360,1.09931299990447,1.27336575629994,1.46109295216759,0.599430485278484,1.05394063585010,0.941602827489124,0.592601491139797,1.41701710369693,0.538877792895886,1.08386909546200,0.883880725334936,1.60471516504847,0.686045855637666,1.08930965396336,0.903916544683015,1.37614733360453,1.15646555509355,0.273308520270242,0.761067967700011,1.56822975092166,1.20340394127572,0.289093629335377,0.344836803540124,1.17463909177071,1.56139590221449,1.82811376542402,1.23348487379153,1.22911294016175,1.80821456807156,1.18275401291265,0.238451668004567,0.840193378994742,1.16583388208111,1.20650195666923,1.71381775908619,0.140060523603003,1.98909165729032,0.670090958814557,1.65720055324280,1.84894471516300,2.13404465033157,2.11620287625837,1.77109772515491,1.71004191373733,1.91503518643479,1.62527976257637,1.94527345539291,2.39547395186389,1.30437635438232,1.33772311830106,1.85530281669854,3.09387777143280,1.35272833628076,0.908327731451414,1.60207275275334,1.29360566126901,0.527674957591285,1.65079553823206,2.06272836141744,0,2.20292316728996,2.39088776981707,1.43619274811433,1.44801227756312,2.00134866724528,2.34836748138111,3.26447563732043,1.67177709211604,1.25984607623924,2.33666311061511,1.79417740598421,2.46053456599870,1.88615337274108,1.46536734229383,1.33517349727627,1.26112677090117,1.71838815593224,2.27398093316464,1.34628580570068,1.71646430987557,1.65018031837875,1.49746114341936,1.41838580790640,1.02216315043597,2.80757684694069,1.52839333209942,2.98432026150105,1.81020704581578,1.46214863223242,1.78591418915000,2.26611521969109,2.69600042473195,1.80893634312078,2.30343326883155,0.852875886901052,1.79384041895374,1.26489271284847,2.12559428401853,1.79077316644999,2.26153006804533,1.43103476132428,2.74789619743659,1.51488555942573,1.55116682320577,0.408637951317614,2.64798577586988,2.50681389977127,2.24401636563271,2.38713850063015,2.06091814641535,2.07475159342855,1.31639392875982,2.33963320313854,2.39212131873294,0,1.10052385648596,2.78512372651713,1.98221848214770,2.02012466138078,1.99005557301148,2.45609102296833,1.85868117165747,1.57784495162619,2.47517015901507,2.07076912196444,2.53253011129188,1.23826846052192,2.20082529543486,2.21676556593255,1.79355419628457,1.89934703802937,2.45310321938973,2.04354648719261,2.40836098941310,2.60150484197839,1.69029933380239,1.50395767152720,2.18887915058641,2.39094945570659,1.66086750416768,1.80288067963751,2.11154763946485,2.66340353052438,3.09751508753510,2.26451865748609,1.79138044934625,1.91644320947097,1.51891335232292,2.66617562847436,1.59466898095570,2.67652000852910,3.08591716343690,1.34728122600827,1.32089423557885,2.27071501348662,2.35738232466420,2.08219441814280,2.17106967263306,3.00931826165894,1.74470990560592,2.20257393832533,1.97336166133204,1.32174179109475,2.56839307394407,2.17281275403546,2.01301843996011,2.34260080389770,0.146747975322957,2.35232974397583,1.38565086763871,0.383109645068135,1.87407651766334,0.861551840826782,1.20865780552570,1.07089251057795,1.67137858029095,1.87322033920855,0.941084301408942,1.42426973649374,1.44305863746221,2.00911592141031,2.51935241505398,2.06576772010144,1.41205362801000,1.34118877492427,1.62222771911686,1.38011871830731,1.24284438165324,1.85250846976837,1.74392623849374,2.21558389828738,1.50414083682996,1.79395535098061,1.69269305206735,1.86833137631155,1.83971866138872,1.10549392220457,2.48722187416199,1.27327187598448,0.909942700700354,1.36245888768209,2.20813934609639,1.19797971449368,1.54723863017461,1.66528394642526,2.01939172591196,2.18346523714210,1.97872457878750,1.35104645119359,1.12937305073343,1.70378910338074,1.60448471640155,1.77905268174570,1.29579934494936,1.20159645916633,2.44834373281424,2.56348726994155,3.35900697986755,1.70370252266350,1.46961923137068,0.940239013218238,1.92744352904483,1.50411510158131,1.17312743271559,2.30981201168036,0.946383092071509,0,1.72836215950537,0.786417918376460,1.71617396837714,1.28528233951975,1.50220900033963,2.33943340810409,1.91295882690256,1.64883704722639,2.65068846138405,2.05792939408787,1.19722868364745,1.71016724467971,1.67995099507177,2.58930308314560,2.95297578301956,2.19212664610596,1.83817446226395,1.64416071392852,1.93415726414386,1.70790907954806,0.953244549585572,0,3.08146130962786,2.40964520819457,2.29900203386304,2.23039564086549,2.67242779797704,2.34413227352198,1.63518258354933,2.58313386049566,2.23323405504293,1.53850908128182,0.729686673749646,1.96365717624841,1.72646922777230,2.51113640167694,1.39233912589269,1.80195983535330,2.89405334604580,2.00781773216166,1.50834096431752,1.96888107079945,1.78431381110032,2.75165330627288,2.64872719206219,1.47115932485011,2.11800221397282,1.81426800296268,1.25374554376111,2.15858525275469,0.906364338521932,1.99954363509952,2.54565895831563,2.45071013869290,1.58835666674533,2.69550810008978,1.74692735121756,2.42196821482898,2.16272125274640,2.09185043926402,2.49810725903595,1.46140773034147,1.57106200706344,2.85728928564644,1.10461791684356,1.41829967267407,2.52884624616733,2.70956702722763,2.55503302445805,0.909857930570132,1.69399745335172,2.29987701272892,0.931754813312521,1.71704633128257,0.791314794418025,0.948223748841126,1.48102027844818,1.46076447083424,0.657244375155349,1.77180773173724,0.787010155525760,0.510819459751152,0.658663498600304,0.686203296494659,0.756572438049522,0.953907082738015,1.49923839536912,0.0361022299239790,1.63655075150572,1.01410621820933,1.28111501089627,1.46800935680623,1.56812588443844,0.895916209210490,0.682713271713519,1.31285741691272,1.06424555637225,0.876597136826472,1.31316183460629,1.73477534302483,1.58204075557261,0.732675439033705,0.336686574872924,2.02491942646367,1.57709830066140,0.712585021144554,0.919753244891707,2.09395178720009,0.938707466323912,1.15939554444597,0.699525037103169,1.20447365818659,2.25218341222787,1.20020807032264,1.10039232904912,1.48512141799945,1.41180444676794,1.20594673504766,0.565751848772066,0.621014939897023,1.52823372703794,1.15386462642172,0.435724355677253,1.13334721240856,0.813844848190165,1.31295690081144,0,1.57250685992703,1.67323402795597,2.27711356723901,1.04344646166530,0.713823175341664,1.42314935803866,1.73979020137460,2.66018068465276,2.74940309008594,2.14620599926245,2.58678004196020,2.41046223535829,0,1.88952501838546,2.19115554663049,3.12432522544903,1.14426650233963,1.11232233378648,1.44583905211643,1.70573779739911,2.38475105523678,1.15487127468897,2.86258253812943,1.39994238229751,1.46507752993557,2.36537543282304,2.78379602306973,2.60653659668346,2.13860006304169,2.30260547645989,1.61073108563557,2.14556625840169,2.29874203567855,1.60181153788816,1.31911226881375,2.44076855262930,1.86231198872822,1.56009888815523,1.30258166036756,1.92739800519990,1.81694207604809,2.62137362428749,1.46125374000929,1.74800313486061,1.93435563988012,1.90725574307380,1.58216086145019,1.78713769892734,2.53745507071868,1.62112694374657,2.86191367694627,2.17798621573832,1.74403602939806,0.877174858258500,2.37401995323110,1.93373368529040,1.74598298302198,1.77243501852199,2.15335802888501,3.22733832339163,2.14596417959563,1.37786355782595,1.66274935131656,2.07865011840863,2.56215237456914,1.60083462725780,2.00491101956083,1.48864313544335,1.40981888857632,1.85487241378284,2.08913972004425,2.16661002600339,1.96138163626677,1.18378262159635,2.72934915742952,1.10977049755797,1.35477072705242,0,1.73778114414290,1.72474559992247,1.47381252845297,2.08421771493259,1.24037271744835,1.93665326036016,1.42456692128037,1.43649257663074,1.50091839417961,1.20196869684948,2.43716135944701,2.38289261772384,2.33904976156014,1.77247663753188,1.58189913300369,1.92311506868607,2.48175534742323,2.13112667078673,2.12088947730697,2.65419227252723,1.09042885982121,1.60709696762681,3.14927911958398,2.31941056264267,1.70487387749472,2.47279280625340,2.73392982215170,2.17201409989100,1.59803605525427,1.38819977516519,1.09093765233800,1.91288205260877,2.14304509396391,2.42875994509349,2.60835857554305,1.13624891013277,2.46212609405796,1.74705009677709,1.91010748946025,1.79722867193360,1.50585007573908,1.46214774473778,2.43747232160515,1.37606488450729,2.51810582450187,2.41384128245790,1.39889833450901,2.07134984901980,2.10399294499857,2.88099064259995,1.47360097838005,1.90531531559933,2.28939738908431,2.56334577748049,1.65351919388951,2.98119099968506,1.60664576830483,1.17384823178063,1.24490628547427,0.751650570519775,1.46575743444389,1.20330222881972,0.606044258866563,1.28295961283663,0.634029664140809,1.23371143387344,1.80405917645605,0.836612882375011,2.67836825276587,2.10985907090167,1.72049259353842,0.428839240057247,1.24399518712263,1.62710719199452,1.45977238508792,1.32847164678192,0.757457310034192,1.49925246745084,0.859084915784978,1.51497751252032,1.10333362158728,2.44888703826213,1.26033580808072,1.62636391511869,0,1.38871686091747,1.92864174345668,0.865286014814600,0.694903199665833,1.65556911769422,1.06629385544108,0.961922032572938,1.05664153347306,2.42260778918905,2.57557291897650,1.62840354593155,2.00048019902956,1.64245833225285,0.925450565000563,1.68867969468470,1.11709358508781,1.95695581256767,0.545274735595505,0.917931162621948,1.12873898328304,1.94478282594175,1.95278409156716,0.689150516589950,0.512005630513324,0.723613291328106,1.61739436559869,1.34548553877217,1.22396703395379,0.803576701663285,1.32144670929415,0.705620009959338,1.51716721366959,1.81635448630500,1.30736318769078,1.87562597613168,2.13628364497982,2.48899582789760,2.12563163716206,2.88711068441436,2.13673435397540,0.963843773688732,1.00955906254829,2.33421930268733,0.796006415262031,1.36662183111514,1.58549841580233,2.57661912222318,1.87154852091257,1.57797591487360,1.93645860451376,1.73685787196160,2.90414048527698,2.12424756561673,1.99899383680558,2.25008121769962,2.45425450913366,1.43597159240215,2.74660387853837,2.48232682953743,2.11216397224442,2.66397161896840,2.12868548244658,1.48708445536362,1.58405446932909,1.74591686831760,1.95959589066118,1.93024696543111,2.09116718626152,2.12326111407979,2.73043450001547,2.90330430674154,2.67597546510656,1.35786085311649,1.63871156483526,2.00110989752239,1.68381181643700,1.79640167282643,0,1.76042375399970,2.55250539115271,3.44889023337176,0.971704230873441,1.71376477099861,2.01851123735620,1.85213853380010,2.57574818303344,2.18955369619585,2.19925274309608,2.18524391120920,2.13453560066887,2.59126885559805,1.70131828994009,1.48024422519293,2.41822268462298,3.06902487387160,2.17860264110116,2.71321919170688,1.49743836991806,2.03866341754225,0.743549836163986,1.16200042907209,2.07512262477760,0.808607256636460,0.448360944147872,1.44662583966814,1.40892341963115,1.49545060455402,0,1.23621738512514,1.18639932827324,0.934074859272949,1.78367457270952,0.780955052619789,1.33217518568721,1.31608897388854,1.78691120342772,1.66050919304357,1.54723889412492,0.572307016082313,1.47746784374060,1.09651302657587,1.35187316106729,0.774069123710452,0.634526836699558,1.97601395569473,1.09800850281241,0.855880004149775,0.556391374737020,0.404971445482915,1.14404038154505,0.882727609596482,1.19780617982274,1.07099249585329,0.456583870422736,1.20488579665828,0.651186659623430,1.27334826968003,0.834796087746561,1.06690146514913,1.98857330800662,1.51331632963144,1.08486925030928,1.84641590705543,0.816902538811038,1.63852393316432,0.999277762566889,1.34036745592946,1.62754804966628,1.58904384938612,1.51804060044322,1.07749546839658,0.901627208055478,1.35776794166474,1.43665196697081,1.79740257698388,1.85164955834330,0.984800775358556,1.57766359644600,0.348288381840908,1.53441812168250,0.669643475228003,1.06630412020915,0.304500365383251,1.04030712359157,1.89620682669713,1.41954262521711,1.34631971529518,1.63233419402328,1.26225921769070,0,0.729810964998771,1.42642339749873,2.57814726684356,0.646312620018453,2.26010338279861,1.88703231640696,1.55037284781943,2.73395855118985,2.21405879314132,0.966174482171585,1.47490536638149,1.38177553357103,1.69787955210326,1.71811841905087,1.47010046232393,0.249349929118593,1.68684646416463,2.37311026704241,2.33837325439000,0.976180440040419,1.77934074131399,1.91569434235502,1.33648012433609,2.72723024192736,0.999322214342151,1.20047498835811,1.17005688095769,2.06319052883283,1.18328417651258,1.18400827324363,1.55215221323205,1.40888630676375,1.30615210544732,1.48970479286504,1.08921433032047,1.10649405470762,1.90744235258472,1.62060343984445,1.40502688183405,0.955340175336313,0.972351411359794,1.64847522671098,1.83570671978520,2.04876062829605,1.06255093430187,1.57488273825534,1.63715938715860,2.13011846667393,0.740500283368808,1.48182089566279,0.843547587497152,1.42578075489948,0.963242871241798,1.36790209004085,0.651290945499652,1.53483561030399,2.54098752580750,1.83932880082186,1.56423497750124,1.69802245218040,1.20326661088528,2.63882295460682,1.53079031394439,1.64876667487849,0.932143662057707,2.31043434324336,1.93235640780790,2.08631998066325,1.94691204948961,1.67188786208476,1.88842833825508,1.37912759292113,1.71601929202924,1.03582128795101,1.68792386461077,1.73025536290602,1.65258755345002,2.17018202857751,1.75748417393771,2.15837098220677,2.31162577070776,2.21831952077600,1.14423187993322,2.87824203473505,2.33217869567617,2.17719181784047,1.86579448988543,1.37130231779116,1.74809705479731,2.77797249583643,2.27750799484211,2.44341812345183,1.25626265192540,0.877004309772926,2.24809398105113,1.43388818591773,1.52256975023025,1.70537147146748,1.89698582144051,2.13691402158747,2.12918766805724,0.994274490013637,1.50529681373115,1.30530908072900,1.98743605879437,1.74847077525989,1.43872240908097,1.92374981548369,1.56553278187535,1.78163666511776,2.56885106512962,2.02488413267788,2.34339297021744,1.66630626722255,1.60397320490209,2.63803830388022,1.48262548235138,2.04337671338669,0,2.20391740322259,2.30154582455237,0.806321524531926,1.08648812952897,0.721846039984271,1.36280058714805,1.02152355641950,2.46930655677513,1.90903819255943,1.15897878216012,1.44078455485298,1.25224023668187,1.17888357697414,1.61978529257554,1.69613436935635,0.946029440931066,1.74383952720030,1.69009300084930,1.53514922692423,0.865108678289597,1.88342220503311,1.89228992160130,1.50232554968488,0.803451531185958,1.83228204572841,1.68719401644289,2.25115486292029,2.09728855567148,1.34543830823181,0.725282012071862,1.97809062501844,1.48155418018518,1.14438496237032,1.20594124789099,0,1.28052276897540,1.62891479825313,2.61043249424993,1.27437758040751,1.29425561143171,1.39863743784741,1.37576406351217,0.818146432043371,1.25491421988823,1.14449802067526,1.84300609832053,2.48297878352360,2.38793083642739,0.273293138563846,1.68205593623413,1.50196089116407,1.62865814258911,1.31992023291342,0.581441167432381,0.515251535187614,0.957061590814881,1.81710205934493,2.70092769686949,1.44526330259255,1.10317739534246,1.63182319789101,2.23090882481444,1.27852520262217,1.48458402650369,1.33058548200104,1.87452255235236,1.21061158991599,1.64915159951206,0.982656174316747,1.83375550414036,0.762856158316758,1.57882694704380,0.791200989074731,1.50819126388969,1.03885843530400,1.59742913743158,1.56864127493672,2.27267456208006,1.39044540217324,1.45894640940698,1.32043796759948,0.701495884716975,1.55888299432955,1.49821330982169,0.958058821540038,1.21143303798246,1.88531245468536,1.64786496088270,0.350396541756013,0.858147518421524,1.05210483735212,1.99091299421019,0.668689198516906,1.96667799218195,1.50984486075688,0.580485150126694,0.392436136300661,0.329591900960146,1.01216066652948,2.38349051395974,1.69287831017318,1.27640368466195,0.975603541576464,1.12852538783261,1.33591529141365,1.49283836368248,0.995389214040917,0.947337300101053,1.49238745374157,2.25829724889797,1.27382651031358,1.06227260766375,1.74331192812587,0,0.608583656591859,1.38806752331877,0.283035360616404,1.81945793389578,1.04548176052030,0.829348068153459,1.10451188106869,0.738998604243535,1.02812190002255,1.42343004895991,1.55804687570067,1.78805556311456,1.93744681262137,1.02536314360163,1.08952340249799,0.893427519634782,1.64790379658460,2.10535225160592,2.42992250129415,0,2.93548494111779,1.95461418212447,2.52187923044635,2.87443173638065,2.48149519259001,2.81456870929117,1.17829940988038,1.82211047954279,1.93841063646936,2.57111735995073,2.85758406601918,1.68866741108867,1.86929215359884,2.21834040531975,1.91164801197823,2.51438053158356,2.82350312152187,1.44967482804571,1.17184730015954,0.981062859165539,1.99825309401908,2.08686377850369,2.55491144245123,2.04436276098947,1.92456809564925,1.42947325443533,2.68104282205896,1.58081502223022,1.67437808952320,2.36652152766560,2.55415847896385,1.89464651635798,2.29536703391267,1.19574366237899,1.40352534990688,1.38900463329640,2.15717305647539,1.58101588166240,2.48702792021237,1.49387563596474,1.66910747575935,2.41858767957914,2.01384566389682,2.47819561549493,1.98983233627424,2.07020067348592,1.87344009969236,2.44636344415818,2.22625525574468,1.51467691397559,2.01105702758648,2.50553983414200,2.90455327977259,2.85084343997474,1.71948461939873,1.62275742736670,2.72711651386636,1.48321788895934,2.22418271284046,2.71692396488717,1.28070113347021,1.36765727124826,1.34944419537501,0.342522827085543,0.740478184843253,1.30543167043554,0.962474329842491,0.240061057188882,0.0718362851568504,1.02537000156826,1.45896913363689,0.630400268937386,1.41590598338676,0.181029528233931,0.571423710644705,0.431243863568461,0.830973391490205,1.08874054399493,0.317232736972081,0.678474949399149,0.641903487777542,0.739673483832581,1.14965413857298,1.17554603848034,0.636144964609942,0.921046263379846,1.31802166982017,0.879605153646065,0.393670455381385,0.826835906687794,1.03441725937615,0.699677973148456,1.52730589939890,1.01799023975851,1.70273439199216,0.517348346115282,0.420643117723143,0.309899859568675,0.554217431925505,0.704872524322284,0.412169184788218,1.27034036501802,0.624366259232392,0.848129280877207,0.892913603689760,0.799171667391402,1.58157714126826,0.612852480452482,0.724838948183151,0,0.755357973138176,1.46218733940778,0.485279687416819,1.05272097471323,0.980853765483349,2.35792315946790,0.140504125414121,1.16363276926017,1.94485230617700,0.439872271399359,0.994082098922005,2.20190786562714,0.801597218813603,0.598002164623733,0.404292823068658,0.913313835646345,1.58479811794984,1.36270980615889,1.37867275790261,0.994625626262474,1.42935702716425,1.84696182530960,0.529871327678265,0.676898145102877,0.364073904040849,1.61490984207268,2.13628812583922,1.14424789044744,1.75845959127726,0.921645271347127,1.80438442204257,0.414115063623233,0.923601609560629,0.591373745061756,1.91306930101307,1.27554148286156,1.17346113030337,0.546109602793456,1.28633429193318,1.90716584781165,0.800487236679492,1.36484097174042,1.77495347245235,1.13563485668067,1.02511378622813,0.470236146820507,1.07397744259246,1.86791138066546,1.01094544042638,1.54380436193012,1.63026942745925,1.04798602000098,1.22561881804699,0.998812198386608,1.01054381707493,1.62535355060250,1.42439188970253,1.39434191324520,1.48802632744877,0.697229001579085,1.04384714997215,2.19410516429329,1.37840750337657,1.22387146489041,1.49914007932641,1.72907851523148,1.23992874623781,1.45134473916681,0.924441550923063,1.80686199165828,0,0.439770275452587,0.206123343771205,1.06833145905963,1.48728661316765,0.312891265849646,1.09975721320541,0.714753346907673,1.79040073594187,1.85783501728042,1.53790807958651,0.586245347869123,0.807471415708517,1.61260492334544,1.21599827125060,1.17004920613595,1.44398224192228,0.419152184277259,0.568285002513345,1.72030941680098,1.66451277670364,0.173925999103796,0.875408532122518,1.70122424418759,1.52666142080131,1.72799779844967,1.97589219454875,0.611703145296733,1.66874944528118,0.143891259727011,0.666794578481290,0.927260429069829,0.943900080923082,0.448048406531703,1.57877710990796,0.495308433414010,1.54043023766765,0.814293053630377,1.42898884649408,0.872823492023073,0.657390174907796,1.39222164613391,1.05038926170119,1.38831823441686,1.80496848906389,2.05682209466577,1.25586213790477,1.23358034228600,0.327297757137621,1.49522066020309,0.943305336692629,0.769776718879573,0.941632689194901,0.907736116148127,1.37441676518815,0.712683428233416,1.95032610869589,0.361940650097237,1.83639786498397,0.402996888459655,0,0.677954659321423,0.105487942208167,0.941985542533964,1.58719898414790,2.24771605046417,0.594493247919282,1.50189242980836,0.979456724462599,1.75894305841036,1.14972616107464,0.330162862582490,0.579420546578004,0,2.37062307272799,2.40191589003041,2.12481834423978,2.76119027195052,0.708273325476749,2.70958637413348,1.75101474527561,1.76095180771940,1.08506492000591,1.32847169122321,2.42382086404622,2.10217537673409,1.57155134267648,1.62643939662981,1.25202791553562,1.75826244060258,1.78914325399685,1.83835067828667,1.71756254796199,2.12228299318036,1.30312996050202,1.85089271303866,1.74769920241109,1.39146566401576,2.28267538131718,1.67085697760920,2.76766270672391,2.23635835925103,2.39788063333551,1.19252579457321,2.35227665808321,2.44020873287201,1.79529162174578,2.17352948233744,2.33822180933624,2.54193176319377,1.25511474659866,2.26475896356675,1.17115800870363,2.26400602999057,1.78182544584637,1.43807535759274,1.88364780169722,2.33505379622709,1.86617233995667,2.14153794902187,2.56914755712098,1.50851827190655,1.09196780351379,2.31157286200635,1.79568628014634,2.94754882910961,1.98287251023033,1.69169499783252,1.89732877464295,3.19477678284540,2.33891065389128,1.48141378020268,2.73542814765779,1.57445777292449,0.799315772099441,2.23852425777566,1.24638614252868,1.77753555040126,0.700596109835987,1.60062008474503,1.76065205403425,1.42296816404256,1.07694301169620,0.725970984348307,1.79796222208591,1.94407611167550,1.29591107745934,0.533150628165037,2.25004861885805,1.78010766519208,1.16720391524939,1.72416103544757,1.37902312063298,1.36571977921046,1.28174266597470,2.99411574635482,1.34853569720643,1.20178254873093,0.959111734635975,2.55972148254330,1.42413546147665,2.07569430982878,1.16644096883359,1.03722017173051,1.86754278120525,1.89274542219554,0.876210414907475,0.810185663951315,1.50705887240972,1.36715162489119,2.53295212853046,2.26000400698626,2.31200495288650,1.10379774729609,0.875669233684528,2.29922235507558,2.44775655941991,1.41154771304859,1.20559702019692,0.719866043245852,2.39839578947982,1.67723549245584,0,1.13677843083580,1.51010678527278,1.77974121381161,1.81110335315743,0.808780900575093,0.759991473542993,1.04869403443053,1.69313781372513,1.49261806609770,1.29591116080076,1.47760528724666,0.975420397924802,1.83404375647450,2.06281651225711,2.11423790853471,1.18849175985559,1.68233894538941,0.786432061311782,1.64802683648810,1.21542285438232,1.40715620175430,1.04506090095705,2.18810534388849,0.650502455411262,0.953566462774337,2.02790568906043,1.21352986510736,0.889806617152773,1.00252196034545,1.28149601007217,1.15041435575677,0.873588495946652,2.10536407094170,1.90660728924220,1.51356149546270,1.07261596248317,2.25484030921644,0.539894303626614,1.81914522152006,0.704594471695150,1.49972313053576,0.745123643121616,1.18026723081228,1.47534630125260,1.91234686425827,0.932727874905179,1.83603285523221,0.634361861129841,0.752076607646324,1.48418032378788,1.21277037019799,1.05001298972016,1.18614149324426,0.760830876846770,0.902608959568995,1.88251312043813,1.64274017962205,0.998080380237647,2.26224353160476,1.54968512198102,1.73246960303248,2.35110065792138,1.93929411319024,1.49125333964237,1.30668215661781,0.388971818254285,0,2.05846727507852,1.27297315549440,1.38358675289450,1.32949153278214,1.41669723988271,0.552466736622944,1.10535108523706,1.59640453007809,2.52345525182343,1.38018295037528,1.52835602653764,1.01170943832532,1.43702481354608,1.27122764328465,1.23873846212688,0.590749844881878,1.18406257368959,2.00340782192954,1.50067514745287,1.28212542615600,1.09723957519002,1.32724970151375,0.813181509826035,0.591685332108808,1.63790338888108,1.53745697277430,0.575873212342963,0.825318248652023,0.666307056421970,1.84854372645142,1.16919346879805,0.966262251909703,0.966017600685202,0.980264827378608,0.391245316478172,0.731991675290149,1.08168631542537,1.71638371771511,0.807729714882810,1.11554386911547,1.81338749522394,1.18969744302643,1.43284626899846,1.13753320841375,1.09746473175290,2.38915316740772,1.30147728301733,0.929413161022771,1.15698418050077,0.500166616483049,1.73936991858071,0.448483911641691,1.80565606392021,2.25437809498369,0.895426066687107,1.02280142055357,2.17044021143345,0.745672299900388,0.773299071107296,0.665521488172014,1.23324549817581,2.12982881428282,0.804894551704565,0,1.28067901355515,0.782468342264193,0.645819986052738,2.06656887458664,1.63529521416276,0.815955918881262,0.911686681937551,1.44543269396037,0.427452456163139,0.937667536414158,0.314039864634651,1.28521349816763,1.20778286255205,0.921417945164446,2.20124775807912,1.23405806505433,1.30536598448769,0.970116096243030,1.31497083986679,1.08111536025299,1.59435941387187,0.885792877023261,1.09213956567241,1.26902430291005,2.23111060959929,0.674738236981697,0.696753458385384,1.47072728275653,1.79787893976939,2.09667366265212,2.04010885969068,1.31366166742908,1.46892508968587,0.872451452789946,1.22529887643978,2.01120169982698,1.10494580803380,1.99540819323506,0.964309500048359,1.75261023590859,0.866009515416937,1.17067580215692,0.942893377149594,1.45765693101039,0.926777877442654,0.923439036677139,1.93467255425384,0.776636418889035,0.855109000542741,1.46839051257345,1.68262748037437,1.81782158099820,1.27626336535899,1.10533697795071,0,1.50746604275319,1.36384228473760,1.86418822744143,1.16342814901187,0.763238138762198,1.51848335618616,1.37459750253710,1.39263741120303,1.49834762530926,1.42892719200345,0.519750035595580,1.62434917711584,0.959148436455128,0.823681045487646,1.09492925153538,0.638875964716400,1.89528685590987,1.02249161286995,1.76722082975270,2.13486199130551,1.47528134016245,1.56427952639159,1.86257278760492,0.955735002837010,1.00099339725991,0.834740585809048,2.28031231360260,0.902383217446487,1.16176752184913,1.74461907925087,2.15844335129380,0.772657847462337,0.884536868013728,1.44914222009164,2.29815232204194,2.05243990290665,0.487066301761997,0.632412484753445,1.28302908869106,0.416098939341490,1.73373223559855,1.28653364944476,1.76470331945943,0.433277254667972,1.95099468587192,1.59124922566808,0.627498610528131,2.03948976491444,2.02704266740658,1.43224257168823,0.709467220419238,1.25398893468637,1.24285591710711,1.32508473507843,1.83297888627131,0.742621932011327,1.14849443854812,0.502485591334080,1.40795914797138,2.07655648323844,0.474922948507663,0.160474325217850,1.18498243808877,1.79830459487522,1.33038481248705,1.10939535975127,1.03419589180779,1.34982598995701,0.961051579555651,0.941090161824567,1.25964218743683,1.05854562136906,0.234143697322062,0.736308664051393,0,0.791609841073064,0.657323973280027,0.871350953913015,1.30462050244781,1.47016925240791,1.06760472518446,1.23892091689721,1.33404958185864,0.881505981075357,1.65122697306000,1.61872221922753,0.833412247243022,0,2.25045979601157,1.89728481525218,1.84734741262966,2.53196828750260,1.95753830045244,2.63868041890823,2.48015427847485,1.81209490119273,2.42729001798511,2.08536983563699,3.26039878036304,2.39684729682468,1.24275541716469,2.27033947477302,1.38620674874809,2.45032833704616,2.22731170322259,1.08758568327701,2.14508328951820,2.72376543881130,2.70392583121636,2.32014485755171,1.39758836814926,2.31742861859698,2.27070013607170,2.16632561090057,2.08320203248854,1.36634979531685,2.10613271556546,2.34890883862667,2.42216371552464,2.45565215355417,3.19756239024488,2.57658750775084,2.23622780613549,2.54093705904084,2.73174160944280,1.36799811594146,1.68291371861021,2.17112113830687,2.03554859287679,2.96234696775911,1.65962228880898,2.68320896661496,2.44843345873185,2.07985690385517,2.51117772750790,1.79780030848492,2.27314188540766,2.55126135351177,1.97493172815535,2.79334946949476,2.62579627573683,1.26208292769927,2.18792934136329,1.84213783019439,2.16209203444770,2.88425079102664,1.95159934494274,2.91739096227183,2.48867768284642,1.88089753115521,2.35612709174967,0.925396409358805,1.05659616763209,1.60794874211488,1.28982788550732,2.92547122940675,1.84966677680182,2.36417796020846,2.22811096407916,2.21790923145101,1.39445368280431,2.39312817530086,1.50521384448328,1.69802466098720,1.14724674366785,0,3.21565146110715,1.78801122439365,2.16327276027862,0.769089308503446,2.13380828732174,2.10434875336226,1.92436681659891,1.69145057776051,2.68364106013429,2.10772302529128,1.68770245236808,1.63970345884143,1.45979937953473,1.47430792619032,1.87129782783172,1.89114081645025,0.810512447768806,1.25382938118379,1.18036149978831,1.49423514563266,2.13231770268056,1.77363342018119,1.07679569967262,2.04825597507814,2.43700203537361,2.60599942191460,1.62819291618895,1.09380526198990,2.61215951300248,1.83542986180145,2.11363250563387,2.02391553288107,1.27342463087235,2.65401209292717,1.84448145445758,0.985108205883445,1.71996133702398,2.30140434844595,1.73460063208270,1.64594415518262,2.41995787296616,1.67973817814970,1.97460238999518,2.00828698157079,1.09596713167234,1.99633234694071,2.11355760953055,1.67212490437595,2.13775482903561,1.80289111310974,1.60454933977218,1.09414031129587,1.08158021314498,2.00780232464405,2.54287546950789,1.68978648006157,1.39685552310014,2.98085552526743,2.53805599564043,2.40782775926844,1.94275931214962,2.73636331021055,2.48173841249766,2.26897298804634,2.51508092868274,1.66131049178040,1.98685678951403,1.55662410380367,1.61238111433224,2.76649655946180,1.03183798014461,1.28272705131336,2.44126903428621,1.68481915872447,1.27687869455471,1.18405730896204,0,2.69512936101551,1.66269954456686,2.08019989529861,1.48548237955766,1.70090739348911,1.11228143832992,1.97095949227947,1.13017565373059,2.39189986165873,1.74133013356168,1.48624213345131,1.99983635483798,0.847605743160108,1.46503259793891,2.44077113506353,1.27234197590072,2.13602986470865,1.25177310592902,1.86692711942185,1.92112404557757,2.04344452824174,1.76727688998371,1.64476057315655,1.98538062342142,1.38212522658475,0.968225069436319,2.54131956587169,2.27520585943664,1.78770604596513,1.48448845956216,1.99410670414072,2.54459383542020,2.19666331147848,1.70264631631866,2.27010206656850,2.44154480222487};
+
+//The reliability and probability data blow should output this decoded (including frozen symbols) sequence succefully
+// 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 53, 0, 0, 0, 62, 0, 55, 8, 53, 0, 61, 41, 36, 5, 50, 24, 54, 0, 0, 30, 23, 55, 21, 4, 61, 49, 59, 25, 3, 30, 19, 57, 54, 1, 33, 51, 28, 30, 24, 39, 38, 17, 41, 27, 54, 34, 45, 37, 35
+
+// this relialbity is for SNR=-10dB
+constexpr uint16_t reliab_seq[GF] = {63, 62, 61, 59, 55, 47, 60, 31, 58, 54, 57, 53, 51, 45, 46, 43, 39, 56, 30, 52, 29, 50, 49, 44, 27, 42, 41, 23, 38, 37, 28, 26, 48, 35, 15, 25, 40, 22, 21, 36, 19, 34, 14, 24, 33, 13, 20, 11, 7, 18, 17, 12, 32, 10, 9, 6, 5, 3, 16, 4, 1, 8, 0, 2};
+
+const value_type chan[GF * N] = {2.62045e-05, 0.00188107, 0.0055512, 0.00859346, 0.000650895, 0.000156559, 4.30483e-05, 4.58585e-05, 1.40256e-06, 9.02546e-05, 0.000178026, 0.000681898, 1.32861e-05, 1.13419e-05, 0.00033873, 0.00394301, 0.00176242, 1.25976e-05, 4.22525e-06, 1.23659e-05, 3.51814e-05, 5.5452e-05, 0.000609942, 0.00115295, 0.00120903, 0.000150344, 0.0174537, 2.23274e-06, 0.000351046, 0.000261074, 0.00384483, 3.30663e-06, 6.3383e-06, 5.79146e-05, 0.00123632, 0.000323354, 5.6716e-05, 0.00304935, 8.99783e-06, 0.00575283, 0.000193345, 0.000586244, 4.91013e-06, 0.000114766, 5.7932e-06, 2.28751e-05, 0.000594688, 1.37526e-05, 0.000215582, 0.000304923, 0.00108942, 4.92504e-05, 0.000992215, 0.0520017, 0.00220255, 1.3247e-05, 1.2427e-05, 1.05189e-05, 3.79476e-06, 3.17689e-05, 0.881808, 8.01779e-06, 9.38767e-05, 7.58717e-06, 7.31022e-07, 0.00060923, 0.000288089, 5.29029e-05, 8.42847e-08, 2.22089e-05, 2.62843e-05, 4.40398e-05, 2.68649e-05, 2.15742e-05, 8.54597e-05, 0.000282119, 0.000285877, 4.77728e-05, 6.05517e-05, 0.00936295, 0.000690204, 0.00076341, 0.00343141, 2.52628e-06, 2.05253e-07, 0.000911277, 0.000237754, 6.12028e-05, 0.0211547, 4.4024e-05, 0.00636689, 0.000846832, 0.00203789, 0.000328541, 0.000186301, 0.000522578, 0.00697983, 0.000677286, 3.46381e-07, 0.000187738, 0.925983, 0.00060606, 0.000181804, 8.25117e-06, 0.00558307, 8.86269e-05, 0.000317969, 3.10237e-06, 0.000247702, 0.000180088, 0.000508028, 0.00123298, 0.00114051, 0.00197455, 0.000560458, 8.28948e-06, 4.67247e-06, 8.27861e-06, 0.000101629, 1.93686e-05, 6.49122e-05, 0.00136917, 2.82682e-05, 1.83211e-05, 0.00272166, 0.000353071, 1.96859e-05, 1.86899e-05, 1.63439e-05, 0.000574466, 0.000754429, 0.000244037, 8.00404e-05, 0.000241489, 9.42531e-06, 0.00103079, 0.00148208, 0.00422754, 5.09506e-05, 0.00114145, 0.000253331, 0.00868304, 0.0180232, 0.010549, 0.0275815, 0.000137058, 0.0330353, 0.00177955, 0.00298548, 2.19367e-05, 0.0141375, 0.00702802, 3.81212e-05, 0.000123972, 0.00204802, 0.000119954, 9.4187e-05, 0.000162059, 1.87445e-05, 4.10513e-06, 0.000299143, 0.0021569, 0.000511373, 0.000723428, 0.00122074, 0.00598556, 0.00658479, 8.14182e-05, 0.000861084, 0.341871, 1.8168e-05, 4.47227e-05, 0.000398206, 0.026186, 0.00229376, 0.000779369, 0.0055419, 0.00329718, 0.000140213, 0.000604184, 1.54988e-05, 0.395251, 0.000656432, 1.03887e-05, 0.000114239, 0.0585281, 0.00655519, 5.39819e-05, 4.79461e-05, 0.00125622, 0.000197197, 0.00103767, 0.00465776, 0.000546513, 0.00630421, 0.000138413, 0.00103064, 0.00020595, 0.000154731, 4.45852e-05, 0.000131217, 7.91924e-06, 0.00280068, 2.45078e-05, 0.000332617, 1.22795e-05, 0.000297769, 0.000336331, 0.0032978, 0.000352624, 0.00620377, 0.0527236, 0.00264069, 0.0166658, 0.00152316, 0.0310376, 2.65193e-06, 0.000367778, 5.9247e-05, 0.120492, 0.128258, 0.0119815, 0.0061499, 0.0406863, 1.58276e-05, 0.117912, 0.000676531, 0.00575599, 0.00542309, 3.53344e-06, 0.00152983, 0.000387257, 0.000108885, 0.000336924, 0.000196165, 0.182108, 0.000344526, 0.00220103, 0.000313011, 0.0289165, 1.2568e-05, 0.000391607, 0.0104022, 0.000770681, 0.00133994, 0.0611113, 0.0562625, 0.0156067, 0.000225571, 6.89287e-06, 0.00087194, 0.0011942, 0.000596786, 0.0649636, 6.21965e-06, 0.000539916, 0.000440496, 0.00324324, 3.57028e-05, 0.000512537, 8.32957e-05, 0.0054498, 0.027903, 0.00521575, 0.0042022, 0.00450934, 0.0182062, 0.00999734, 0.00268917, 0.001005, 0.000141709, 0.0181912, 5.10601e-07, 0.000591128, 0.00362583, 0.000340365, 4.83342e-05, 0.0913572, 0.000279494, 0.000536187, 0.00172051, 0.00149849, 0.00167773, 3.11843e-05, 0.000105397, 0.000306543, 0.000218156, 0.00276309, 9.14865e-06, 0.000300507, 0.000911009, 0.00017644, 0.130954, 0.00472092, 7.20015e-05, 1.72337e-05, 0.000109558, 0.00239216, 0.0473993, 0.0335402, 0.000185268, 0.00765899, 0.103165, 0.00333567, 0.0620038, 0.260168, 0.0731331, 0.000273413, 0.000898665, 9.26998e-05, 0.00634485, 3.21703e-05, 0.00998791, 0.00264623, 5.58848e-05, 0.00702046, 0.000797726, 9.02248e-07, 0.0260779, 0.00859284, 0.24754, 2.47533e-06, 5.69052e-07, 0.000121027, 0.000798467, 0.000267252, 0.121953, 0.000284633, 0.000337275, 0.00058358, 0.0095894, 0.0410832, 0.00114344, 3.35114e-05, 0.0451393, 0.00179957, 0.0102156, 0.00550717, 1.79625e-05, 0.00339289, 2.17438e-05, 0.038295, 0.000663586, 0.000167174, 0.000267084, 0.0721848, 0.00441167, 0.000589267, 0.000128754, 0.000151447, 0.000347407, 0.0235419, 0.000167891, 0.0151541, 1.11159e-06, 0.0274209, 0.000674631, 0.000973175, 0.00609934, 0.0857182, 0.00166894, 0.0020299, 0.000168293, 0.00312986, 0.00562559, 0.000261795, 0.00672635, 0.0051663, 0.000393852, 0.0372024, 0.000112268, 0.00112758, 0.0321655, 0.000936234, 0.00023312, 0.00957834, 0.0313921, 0.00251796, 2.79168e-05, 0.00256867, 0.00537592, 0.0026604, 0.00192905, 0.0802118, 8.52426e-05, 0.000390933, 3.73966e-07, 0.000127946, 1.27801e-06, 0.00130025, 1.29312e-06, 6.40746e-07, 0.00719091, 1.67778e-05, 1.04278e-06, 3.14444e-08, 1.15237e-07, 0.000665213, 1.00904e-05, 2.07044e-07, 3.90808e-06, 5.62489e-06, 2.00445e-06, 6.13534e-07, 8.29974e-07, 5.10183e-05, 1.07404e-07, 0.982528, 0.000132173, 0.000117072, 3.93152e-07, 0.000282323, 6.50356e-05, 3.59404e-05, 9.4003e-06, 3.88309e-06, 4.85936e-06, 1.3808e-05, 2.69387e-05, 1.65917e-07, 0.000144557, 1.15117e-06, 1.80825e-05, 4.3208e-05, 3.97634e-07, 2.36551e-07, 2.53175e-07, 4.12228e-05, 1.30363e-07, 7.58174e-07, 2.8767e-05, 1.5311e-05, 7.51507e-07, 1.7027e-05, 2.69569e-06, 4.4088e-06, 2.43285e-06, 4.18884e-05, 0.00149403, 1.88038e-07, 3.03612e-05, 8.29035e-07, 4.4e-05, 1.43407e-05, 5.54272e-06, 4.91191e-06, 0.00496656, 1.58496e-07, 3.46066e-06, 1.44164e-06, 0.00108855, 1.60267e-05, 6.70741e-06, 0.00158167, 8.35413e-07, 0.000339438, 0.00217876, 4.99391e-05, 0.00147342, 6.67756e-05, 0.000434701, 0.00112378, 0.00407297, 1.59705e-05, 4.53103e-06, 0.000305846, 4.40105e-05, 2.39845e-05, 3.04672e-05, 3.6805e-05, 9.74289e-06, 1.01519e-05, 0.00110178, 8.7686e-07, 6.71771e-05, 3.1619e-05, 0.000124516, 4.64717e-05, 4.50884e-05, 7.78087e-05, 2.00245e-05, 1.42222e-05, 1.96923e-05, 9.6775e-05, 3.49801e-05, 1.33556e-06, 0.000155942, 1.10285e-05, 4.87684e-07, 0.000156268, 7.07371e-06, 0.000665074, 0.000527536, 1.22806e-05, 4.94998e-06, 1.20062e-05, 3.05261e-05, 0.000218859, 5.12784e-06, 4.30621e-06, 3.15423e-05, 3.18285e-06, 0.00302442, 9.71468e-07, 0.980006, 9.53555e-06, 2.60079e-05, 2.05833e-06, 1.26474e-05, 1.33115e-05, 0.000343183, 0.000113233, 0.0283802, 1.14857e-05, 0.00469868, 1.07974e-05, 0.0732614, 0.000687891, 1.86633e-06, 0.000247665, 0.00139576, 0.00650517, 0.0300517, 0.0020038, 0.000633463, 0.0244823, 0.345991, 0.000536797, 5.69346e-06, 0.00270837, 0.00234676, 0.0276408, 0.00609613, 0.00246615, 0.00343598, 0.0100993, 0.0139951, 0.00396535, 3.83089e-05, 1.29592e-06, 0.00198184, 0.00392829, 0.0571687, 0.000327929, 0.00197762, 0.00137343, 0.000671561, 0.00438984, 0.000690893, 0.00129848, 4.02726e-05, 0.0338672, 0.00971447, 0.000353909, 0.00267561, 0.00056489, 0.00440407, 0.00438347, 4.69246e-05, 0.000863875, 0.000167584, 0.00420972, 0.0233325, 0.000709865, 0.000310105, 0.0506436, 0.00048445, 0.000535248, 4.90259e-05, 1.91685e-05, 9.10344e-05, 0.00383895, 3.91801e-05, 5.63285e-05, 0.00294892, 0.190142, 0.000806228, 0.000733829, 0.000846666, 0.000101784, 0.000172191, 0.00318991, 4.39926e-05, 0.0004014, 0.000752425, 0.000116158, 0.0881227, 0.0110706, 0.000362309, 0.000158168, 0.00101655, 0.0135786, 0.0144097, 0.00015695, 0.00410581, 0.000688023, 0.0370423, 0.000214754, 0.00163137, 0.000644624, 0.00772201, 0.00019773, 0.0210262, 0.00748456, 0.00255916, 0.000378519, 7.05596e-06, 0.00207212, 5.73039e-06, 6.38421e-05, 0.00286134, 2.61253e-05, 0.0315899, 0.0687867, 1.66159e-05, 0.00130088, 0.0205069, 2.60773e-05, 0.000463651, 0.000174178, 6.79772e-05, 0.266687, 0.000685588, 0.00124971, 0.000542287, 0.00520231, 0.00337697, 0.290754, 0.000558941, 0.00149099, 0.000465732, 0.00259385, 0.000457402, 0.0002608, 0.00223376, 6.78032e-05, 0.0743974, 0.000195465, 6.46437e-05, 0.00101188, 0.00128932, 0.000943476, 0.00128169, 2.40506e-05, 0.00131161, 0.000152939, 0.000249758, 0.000700906, 0.00604987, 0.0148785, 1.37307e-05, 0.000670178, 0.000662547, 5.53777e-05, 0.00795734, 0.000285578, 0.0129043, 0.417668, 0.00755939, 0.000796963, 0.0235506, 0.0261219, 0.000162246, 2.50336e-05, 0.000348035, 0.00199306, 0.0228394, 0.0233222, 0.000198454, 0.0366086, 0.164591, 4.61707e-05, 1.83106e-05, 0.000245194, 4.56266e-05, 0.0143991, 0.000300189, 0.00066692, 0.00107712, 0.00488741, 0.118782, 0.00220238, 0.00096269, 0.00137545, 0.00453677, 0.000489485, 5.31149e-05, 8.63644e-05, 0.000476975, 0.00185397, 0.000185322, 0.000567889, 7.56509e-06, 0.0182186, 7.44549e-05, 0.0430621, 0.00105486, 0.000114063, 0.000364296, 0.00255053, 0.00135914, 0.000116613, 0.00459956, 2.64397e-06, 0.0012896, 0.000255878, 0.000493269, 0.000623096, 0.000493309, 0.000587332, 0.00362437, 0.747729, 0.000713554, 0.00124563, 0.0121861, 1.60005e-05, 7.35279e-06, 0.00369712, 0.000558205, 0.00396328, 0.00145211, 0.000493217, 0.00272544, 0.0027373, 0.000238824, 0.000454504, 0.000115542, 0.000164009, 0.00361665, 4.28854e-07, 5.45158e-05, 0.000953426, 0.000740036, 0.000725934, 0.000756199, 0.00181851, 0.00037591, 0.0212257, 1.6999e-05, 0.000606396, 0.000734371, 0.0381783, 0.000135352, 0.00389194, 8.10604e-06, 7.94555e-05, 0.00342128, 0.00930001, 2.74058e-05, 0.000604018, 0.0397236, 0.00123244, 0.000501963, 0.000101317, 0.000108943, 0.0457313, 0.024272, 2.57371e-05, 0.000805689, 4.9188e-05, 0.00184903, 0.00825158, 0.00124253, 0.000115737, 0.0016768, 0.000741258, 2.31881e-05, 0.000412883, 0.00683303, 0.00201216, 0.000217894, 5.8125e-05, 0.000714136, 1.93889e-05, 0.00658223, 0.000117293, 0.000368336, 5.09699e-05, 0.00366617, 0.000587548, 0.00122476, 0.00453068, 0.000659779, 0.000735229, 0.0289092, 0.0139075, 0.0519842, 0.0104072, 1.59746e-05, 7.71451e-05, 2.34507e-05, 0.000329745, 1.49501e-05, 0.0145372, 0.0221293, 0.000413675, 0.0049242, 0.00267153, 0.168122, 0.0130634, 0.0208566, 0.00755165, 5.36829e-05, 0.000548188, 0.00189417, 4.78067e-05, 0.000141519, 0.218787, 0.0167016, 0.0292291, 0.000193676, 0.000595677, 0.274767, 2.58589e-05, 4.2159e-06, 2.91246e-06, 6.67962e-06, 0.00141221, 0.00088997, 0.000185164, 1.27996e-05, 7.6888e-05, 0.0547802, 0.00010048, 0.00035072, 0.00143268, 0.000328291, 0.00121674, 6.60468e-05, 0.000130652, 0.00671489, 0.000987243, 2.86249e-06, 9.62966e-07, 4.34168e-06, 0.00349245, 2.15564e-06, 6.15274e-08, 2.29683e-05, 3.49057e-05, 0.000655497, 1.1959e-06, 4.42856e-06, 1.19764e-07, 3.90739e-05, 0.000190376, 0.000122569, 5.19971e-06, 1.43359e-05, 1.27383e-05, 1.56067e-05, 1.39918e-05, 4.64974e-09, 2.39616e-05, 0.000280676, 4.94184e-05, 2.14884e-06, 1.61907e-06, 0.991419, 0.000447415, 0.000167073, 4.23625e-05, 5.55129e-06, 0.000866941, 1.25283e-05, 1.21172e-07, 1.49433e-07, 1.59569e-05, 0.000115495, 0.000256994, 1.80339e-06, 1.14902e-06, 1.01831e-06, 0.000321943, 9.09604e-06, 3.67153e-07, 4.50823e-07, 1.8561e-05, 1.8903e-05, 3.16307e-07, 2.43699e-07, 1.96533e-06, 0.00025476, 5.60685e-05, 4.81689e-05, 0.000105446, 1.45484e-06, 4.74212e-07, 1.22628e-05, 0.000522567, 1.43271e-06, 4.34086e-05, 1.19292e-06, 2.53918e-05, 0.000206795, 1.22246e-06, 0.000366562, 2.16371e-05, 5.15215e-05, 2.07505e-05, 0.000144078, 0.000340494, 2.8793e-05, 7.10353e-05, 0.00169001, 0.00233073, 0.00316862, 3.83311e-05, 0.000217616, 0.000118953, 0.000217544, 1.98525e-06, 3.49372e-06, 3.71398e-05, 0.951624, 7.61239e-05, 4.36124e-05, 4.80824e-06, 3.18008e-06, 0.0296021, 3.0818e-05, 5.67512e-06, 0.000191957, 2.69906e-05, 0.00020345, 8.7244e-05, 5.70361e-05, 1.04213e-05, 1.16228e-05, 0.000197033, 7.85622e-06, 3.59927e-06, 0.00230481, 1.00958e-05, 0.000195979, 1.46843e-05, 4.81383e-06, 0.000102701, 0.000138964, 8.72285e-06, 5.51763e-05, 5.59116e-05, 0.000274021, 0.000164877, 5.10257e-07, 1.11224e-05, 0.000424202, 2.40122e-06, 3.09535e-05, 0.000452596, 0.000372126, 5.83522e-05, 0.000119677, 5.21402e-05, 5.75957e-05, 0.00159953, 1.5245e-06, 6.92313e-05, 0.00219313, 0.000167596, 0.000308182, 4.24683e-05, 2.90767e-07, 0.000567927, 4.93932e-05, 1.88801e-05, 1.23042e-05, 0.0012236, 1.92996e-06, 0.000200773, 1.82823e-06, 0.000677459, 6.14555e-06, 0.000145058, 0.000162831, 3.94005e-06, 9.45181e-07, 0.00017641, 2.2014e-06, 7.30031e-06, 6.86927e-06, 3.81305e-05, 0.828455, 0.00270021, 1.57388e-05, 4.31261e-05, 3.08737e-06, 0.000324255, 0.00375943, 0.00759025, 3.33346e-06, 0.000107061, 1.30199e-06, 7.67357e-06, 2.73413e-05, 1.04403e-06, 0.138758, 0.00779196, 0.000430146, 3.80524e-05, 0.000305668, 0.000984775, 5.31797e-06, 2.01193e-05, 2.04742e-05, 0.00373271, 5.91883e-05, 1.78433e-05, 0.000110278, 8.71753e-05, 0.000149071, 3.39281e-05, 9.03742e-05, 9.08941e-06, 2.43054e-06, 8.00524e-05, 5.57347e-05, 1.78769e-06, 2.0753e-07, 2.87278e-06, 0.000411314, 1.28924e-05, 3.49327e-05, 5.92418e-05, 0.000106776, 0.000988805, 2.29027e-05, 0.00103095, 0.00562847, 0.00163939, 0.00559384, 0.00600114, 0.113588, 0.303006, 0.000219089, 8.09219e-05, 2.98326e-05, 6.67834e-05, 9.97164e-06, 2.45241e-05, 2.20575e-05, 6.32664e-05, 0.00088746, 0.000222004, 0.000304433, 0.000220944, 0.000551612, 0.000128928, 2.991e-06, 0.0024976, 0.000189411, 0.000917398, 6.59011e-05, 0.000347858, 1.37385e-05, 3.41984e-05, 2.40974e-06, 3.16969e-05, 0.000466097, 0.00275022, 1.7016e-06, 0.000302567, 7.30388e-06, 8.68765e-05, 0.0117996, 0.000106804, 7.11767e-06, 7.90584e-06, 0.00209942, 0.000812757, 0.00021049, 1.25749e-06, 0.000282235, 0.00188911, 0.000252445, 0.000101364, 0.0542126, 0.000693399, 0.000508323, 2.58214e-07, 2.35505e-05, 9.31552e-07, 0.00883786, 0.41948, 0.0003964, 2.50027e-05, 0.0500614, 3.40251e-05, 0.00154678, 0.000153704, 0.000112573, 0.00164511, 1.50376e-05, 0.025482, 4.84783e-05, 0.000237233, 6.36084e-05, 2.47945e-05, 0.000183158, 0.140243, 0.00295218, 0.0053879, 0.0269424, 0.000128987, 0.00013066, 0.00105122, 0.000359763, 0.000163385, 9.32536e-05, 0.0384512, 0.0224082, 0.00122273, 6.30679e-05, 0.00131277, 0.000297312, 0.0160016, 0.00458723, 0.01566, 0.000914777, 0.000121622, 0.000182275, 0.000534822, 7.73883e-06, 0.00974895, 0.00299535, 2.39485e-05, 0.000177142, 3.70635e-05, 0.000838567, 0.00166013, 0.000123359, 0.305731, 0.000329569, 0.0020865, 5.61239e-05, 0.000161274, 1.40861e-05, 0.0476349, 0.00847668, 0.294724, 0.000156237, 0.000328871, 0.0015804, 5.67049e-05, 0.000502207, 0.000282853, 0.00106389, 0.000461193, 1.07404e-05, 0.00589312, 0.00300796, 0.00310657, 4.99256e-05, 0.000373312, 0.0139217, 8.43295e-05, 0.00100485, 0.000225911, 3.97584e-05, 0.000493845, 0.00302322, 0.0665882, 0.00200081, 0.00124109, 5.32124e-05, 0.0326841, 0.000727387, 0.00369669, 3.98103e-05, 0.00683145, 0.00040012, 0.029468, 3.14777e-05, 0.000758425, 0.00577806, 0.0181184, 0.00112792, 0.000464924, 0.00014311, 0.00147965, 0.0428218, 0.00426629, 0.00890362, 0.000760665, 0.000174597, 4.78028e-05, 0.0062396, 4.50104e-07, 6.69448e-06, 0.667817, 0.000100207, 7.72809e-06, 5.62256e-05, 0.00126788, 0.000527432, 0.00677529, 0.000157081, 0.00556483, 7.44935e-05, 1.88851e-05, 3.71167e-05, 3.9006e-05, 0.00309945, 0.0101618, 0.0173945, 5.23895e-05, 0.00028002, 0.0208938, 0.000461445, 0.000255383, 0.000401918, 0.000532143, 0.000638647, 2.49111e-05, 0.00079113, 0.0084981, 0.0198306, 0.00232712, 0.0724143, 0.000108317, 8.33055e-06, 9.54595e-05, 0.00556651, 3.29405e-05, 0.00187975, 0.000910437, 0.0150851, 0.00657271, 0.00655181, 0.0246667, 0.0576855, 6.60728e-05, 1.40424e-05, 0.000115724, 0.000608117, 3.96167e-06, 0.000204606, 0.10935, 0.000747005, 0.00705409, 0.000381634, 0.000318323, 0.000454013, 0.0168743, 0.00413658, 0.000721401, 0.00175224, 0.000181027, 0.000220614, 0.000489737, 5.65517e-05, 0.034844, 0.00985266, 0.0014041, 3.57551e-05, 0.000115392, 0.000609067, 0.0020666, 0.000251919, 0.000281592, 0.0149444, 0.00700147, 0.000640325, 3.81854e-05, 2.83831e-05, 0.0754935, 0.0280816, 0.00112063, 0.000166816, 1.44583e-05, 0.0804419, 0.000403693, 0.00897015, 0.000902394, 0.000211102, 0.000152134, 0.0607601, 0.000122493, 0.313499, 5.99756e-05, 0.00882394, 0.0027679, 0.00171359, 0.00158686, 0.000157591, 2.51066e-05, 0.000735035, 0.000683981, 0.00612555, 0.000544151, 6.82428e-05, 0.00153376, 0.00976615, 0.00798499, 0.241448, 0.00867389, 0.0032019, 0.00156053, 0.000568714, 5.76549e-05, 0.000768705, 0.00692051, 4.10569e-05, 0.0048831, 0.0897319, 0.00350534, 0.000676932, 0.0139215, 4.32329e-06, 0.130074, 0.000526234, 0.000122982, 6.40214e-05, 0.0272538, 4.44038e-05, 0.00214186, 6.81602e-05, 0.00125283, 0.169281, 0.000264025, 0.00328901, 0.0133441, 0.000831483, 0.000977881, 0.000196443, 0.000333851, 0.0486286, 0.000344614, 0.000202939, 4.57132e-05, 0.108023, 0.000272973, 0.0182618, 0.000355453, 0.0513818, 6.12408e-05, 0.000818268, 0.000184345, 0.000461416, 0.000567251, 0.000161082, 0.000129489, 0.00151408, 3.97627e-05, 0.00254579, 0.0756007, 0.000295926, 0.00218463, 0.00193327, 0.00470974, 0.0217414, 6.85579e-06, 2.66957e-08, 0.002191, 3.77864e-05, 0.0730046, 0.000490484, 0.0965613, 0.269672, 0.0001127, 0.000310326, 0.0228651, 0.0144605, 5.9557e-05, 0.000396717, 1.95444e-06, 1.92209e-05, 2.12841e-06, 0.000303364, 0.0013152, 0.00748368, 0.000180894, 0.00151807, 1.72456e-05, 3.47823e-06, 8.14478e-05, 0.00217451, 0.000940147, 1.72064e-05, 0.000184004, 0.0126996, 0.0853603, 0.000686728, 0.000104193, 0.0203379, 0.00789298, 0.000830353, 0.00137212, 0.000168755, 0.0189434, 0.0382081, 0.00333667, 0.0020749, 0.00249741, 0.00176734, 0.000658144, 0.00025958, 0.0337536, 0.11739, 6.76345e-06, 0.0276006, 4.90015e-05, 1.60544e-05, 1.81597e-05, 0.00866713, 1.52263e-05, 0.0118076, 5.41075e-05, 0.00621648, 0.0117897, 0.0235271, 0.000438404, 0.00132308, 0.00476758, 0.00157379, 0.00035145, 0.00353865, 0.000273265, 0.0255059, 0.00717452, 0.011102, 0.0173702, 0.000277468, 0.00211385, 0.000182221, 0.00750354, 0.000510883, 5.50129e-05, 0.271375, 0.0429777, 0.0707143, 0.00346538, 0.000111183, 0.000420123, 0.0612694, 6.5414e-05, 0.0106095, 0.000141551, 0.0017338, 0.000909629, 0.211611, 0.000559908, 0.00241011, 0.00892013, 0.00610947, 0.000480337, 7.76885e-05, 0.000566945, 0.000222367, 0.0120528, 0.00568784, 0.000534339, 0.00246877, 0.0674776, 0.00206567, 2.19924e-05, 0.00303374, 0.00185376, 0.00926113, 0.000181957, 5.69812e-05, 0.0266624, 0.0219518, 0.00710641, 0.00746821, 0.00411091, 7.37879e-05, 0.000146395, 0.00293863, 2.59125e-05, 0.0022447, 0.0022284, 0.00591734, 0.000299496, 0.000679247, 0.000503322, 6.58963e-05, 0.000131004, 0.786468, 0.000484537, 0.00567364, 0.000843037, 0.000120431, 0.0202258, 0.000374361, 0.00199301, 3.00384e-05, 0.000399031, 0.000710651, 0.00453916, 5.98473e-06, 0.00120886, 0.00144392, 0.000362685, 5.11201e-06, 0.000411037, 1.22856e-05, 5.78533e-06, 0.000309852, 8.02797e-06, 4.12627e-06, 0.0243546, 0.00010245, 0.0130975, 0.0627999, 0.000219695, 0.00147549, 0.0245269, 0.00310991, 1.53123e-05, 9.37463e-05, 0.000769328, 1.01931e-05, 0.000856593, 0.00774628, 0.000696605, 0.000125837, 4.08979e-05, 0.000256637, 0.00257901, 0.0141176, 0.000144056, 0.000165531, 0.000206669, 0.000384758, 9.94228e-07, 2.38878e-05, 0.00113484, 4.8915e-05, 0.000616311, 6.84289e-05, 0.00130752, 1.05721e-05, 9.10178e-05, 0.00552825, 3.76835e-05, 6.78483e-05, 0.000420641, 0.000668087, 5.28464e-05, 0.00182909, 0.000303381, 0.000342008, 4.24117e-06, 7.44529e-06, 0.0204144, 0.000791402, 1.73075e-05, 1.17214e-06, 0.00598216, 0.00686475, 0.00876001, 5.08813e-06, 8.90285e-06, 0.000133089, 7.90789e-05, 0.00115192, 0.000249628, 0.000155836, 0.00251287, 0.0131197, 0.0202769, 0.000288038, 0.00065834, 0.000322575, 6.93728e-06, 4.71694e-05, 2.1429e-05, 2.99578e-05, 0.0173635, 0.0277269, 0.000713428, 3.55272e-05, 8.08629e-06, 0.00176607, 0.000180342, 2.7941e-06, 2.08265e-07, 0.0350069, 0.000838989, 0.00482754, 2.25039e-05, 0.556602, 9.336e-06, 0.0798256, 0.00196609, 0.000267538, 0.000199343, 0.000988293, 0.163575, 0.00269931, 1.4925e-05, 0.000599312, 0.00966206, 0.00429167, 0.00503636, 2.44953e-06, 3.10376e-05, 0.000135601, 7.50307e-06, 0.000125393, 0.0117044, 0.00290644, 0.000867449, 0.000623323, 0.0203351, 0.00242133, 0.00607755, 0.254515, 8.60316e-06, 0.000142107, 4.62453e-05, 0.27454, 0.0338904, 0.000202892, 0.00288822, 0.00423927, 1.50055e-05, 0.00304838, 0.000533665, 0.00105319, 2.38956e-05, 0.00120246, 0.0010538, 1.37691e-05, 6.98382e-05, 0.00136732, 0.0031332, 0.000445152, 0.0029564, 0.00166241, 0.00193946, 0.008852, 0.000340172, 0.000298148, 0.018018, 0.0151998, 0.0794168, 3.31985e-06, 0.000536711, 0.000359046, 0.00295155, 0.00484857, 0.0108884, 0.0121814, 0.0397924, 0.0265465, 0.000895024, 0.0240706, 0.000791554, 1.7394e-06, 0.000345022, 0.000324031, 0.0146976, 0.00193588, 0.00372126, 0.00146978, 0.075741, 0.00873917, 2.14421e-05, 0.00038309, 0.000365369, 0.0106377, 0.00157524, 0.00253729, 7.41576e-06, 0.0142893, 0.00205627, 0.000120747, 1.37899e-05, 0.000165228, 1.51062e-06, 0.919896, 0.0128943, 0.00447411, 2.44668e-05, 7.66135e-07, 2.17153e-06, 0.000618425, 3.33266e-05, 0.000325125, 1.28292e-05, 0.000717867, 4.37321e-07, 2.58869e-06, 4.84248e-05, 1.29959e-05, 1.15718e-05, 1.28049e-05, 5.12312e-05, 0.000231554, 5.65213e-06, 0.000797481, 0.000248774, 0.00241594, 0.000648613, 0.000311715, 2.78394e-05, 4.55707e-05, 0.00165907, 3.94133e-05, 0.000486676, 0.00100684, 0.000591737, 0.0163768, 0.000891928, 0.000155648, 4.59525e-05, 3.50724e-05, 0.00017817, 0.00172662, 5.29584e-05, 0.00037576, 0.00171145, 0.0034676, 9.24977e-05, 3.84154e-05, 0.000512462, 6.5924e-07, 1.53056e-07, 2.9394e-05, 1.89534e-05, 0.00435919, 4.51008e-06, 0.000156744, 0.000384173, 5.79647e-05, 0.0024785, 0.0510788, 0.000554608, 0.000279932, 0.00500511, 0.000610148, 0.00630051, 0.000760565, 0.315013, 0.000576789, 0.0023717, 0.0422372, 0.00137453, 6.01016e-05, 0.0649856, 0.00657804, 0.000121738, 0.00028466, 4.80416e-05, 0.000109277, 1.37773e-05, 0.000296957, 0.00214992, 0.0482381, 0.00372534, 4.10297e-06, 0.0239886, 0.000664965, 7.6213e-05, 1.07772e-06, 0.121311, 0.0136538, 0.000492967, 2.70346e-05, 5.98387e-05, 9.83267e-06, 0.0383867, 0.000666007, 0.000219809, 0.00172763, 0.0017913, 0.0225648, 6.91752e-05, 9.10318e-05, 0.0158608, 0.0031096, 0.000862517, 2.49415e-05, 5.8765e-05, 0.00246448, 0.0792966, 0.00286541, 0.0018275, 0.00193886, 0.00642311, 0.0169588, 0.0121604, 0.000989869, 2.01608e-05, 0.00669566, 0.0590181, 0.00388505, 0.000127735, 0.00166184, 0.00516914, 8.0081e-05, 0.00267182, 0.00214844, 0.0370967, 0.000973367, 9.85355e-05, 0.00516989, 0.0088639, 0.00016396, 0.00124469, 0.00259198, 0.0139893, 0.00272691, 0.00417194, 0.000138889, 0.000804503, 0.140705, 0.00736645, 0.00028754, 3.76418e-05, 0.00084482, 0.000131951, 0.00112223, 0.000617401, 0.000146319, 3.30131e-05, 0.00269244, 0.0861138, 2.97815e-05, 0.00125997, 0.000559973, 0.0076788, 0.00106793, 0.00423457, 3.8416e-05, 0.00331302, 0.00811634, 0.0109246, 0.0016327, 0.000224277, 0.00781619, 0.000106112, 0.0319667, 0.00272195, 0.220932, 0.0228528, 0.00116382, 0.000535388, 0.00600587, 0.0012166, 0.195858, 0.00937498, 0.00034146, 9.35109e-06, 0.000148679, 0.00095342, 0.00353157, 0.00165643, 0.00837161, 9.39484e-05, 0.00217349, 0.00176621, 0.11823, 5.90039e-05, 0.0228954, 0.021619, 0.00752884, 0.000111503, 0.0101333, 0.00188123, 0.0125258, 0.00936221, 0.000434038, 0.000169898, 0.191721, 0.0171243, 0.00365924, 0.00125636, 0.0016596, 0.0465928, 0.00109839, 0.00481411, 0.00162839, 0.0716003, 0.0410259, 0.000555694, 7.27023e-05, 0.00107403, 0.00123473, 2.03257e-05, 0.00735395, 0.000491949, 0.00896343, 0.0039229, 0.0110587, 0.00129267, 0.000337953, 0.0335899, 0.020901, 0.0141547, 0.00111641, 0.00190896, 0.0428131, 0.00159223, 0.00526727, 0.000163818, 0.00867641, 0.0206794, 0.0355639, 0.000725968, 0.00686647, 0.00439666, 7.22815e-05, 0.0124637, 0.00022103, 0.00256529, 0.0127671, 1.271e-05, 0.00116005, 0.163018, 0.00115566, 0.00126339, 0.0611763, 0.0253355, 2.66955e-05, 0.0031038, 0.0119986, 2.33436e-05, 0.00474201, 0.00312662, 3.1894e-07, 6.63616e-05, 0.000455765, 3.46959e-05, 3.74844e-05, 0.00096617, 1.41003e-05, 0.000113195, 0.000187005, 0.00449205, 2.50389e-05, 0.000239962, 0.00203124, 0.0127839, 2.58172e-06, 0.00104311, 0.000105679, 0.0162835, 0.000666458, 0.000950942, 1.77252e-06, 0.000260128, 4.63061e-05, 0.00228464, 0.0016599, 0.000158863, 0.00248474, 0.404221, 6.29556e-05, 0.000952469, 0.00261277, 0.00042131, 0.00141261, 0.000500585, 0.0996663, 6.24961e-05, 0.000153948, 0.00279201, 0.266258, 0.000818559, 0.00494331, 4.42593e-05, 1.99852e-05, 6.4409e-05, 0.000639094, 0.000163452, 0.000409505, 2.26713e-05, 3.50336e-05, 0.0280973, 3.74969e-06, 0.00084677, 0.000554826, 0.0649017, 2.36406e-05, 4.60273e-05, 0.0130991, 0.000173172, 0.0317705, 0.000171089, 0.0186095, 0.000161057, 3.0522e-05, 0.0595669, 9.1428e-05, 1.38715e-05, 7.54946e-05, 0.00112975, 0.000110184, 0.00240819, 0.00105306, 1.83443e-06, 0.000131652, 0.00125958, 0.000601452, 0.000787621, 0.000873793, 0.000213707, 0.0269691, 0.000405159, 6.66005e-06, 8.46976e-05, 0.000712803, 0.555353, 0.00715176, 8.53215e-07, 0.00254571, 0.112384, 0.000241028, 0.000109022, 2.79224e-05, 0.000212388, 0.00191291, 5.55483e-05, 0.00874026, 0.000803399, 0.000188752, 0.0020232, 0.000665085, 0.0146365, 0.00370269, 0.000286981, 0.0973698, 3.52482e-05, 0.00474345, 4.93587e-05, 0.0228316, 0.000283064, 3.56274e-06, 0.000258518, 0.00321923, 0.000298388, 0.0569755, 4.29506e-05, 0.000917407, 0.000765044, 2.80701e-05, 1.21357e-05, 5.06184e-05, 0.00247915, 0.001127, 0.00012644, 6.63886e-05, 0.000244996, 6.25257e-05, 0.000441243, 0.00369001, 0.000452309, 0.00126377, 0.000888417, 0.00186255, 0.000334291, 0.00314932, 0.000464117, 0.000653434, 0.0206712, 0.0105593, 0.000466136, 0.0147206, 0.004554, 0.000219935, 0.000286767, 0.000202191, 0.000888746, 8.09219e-06, 8.74688e-06, 0.357438, 0.00115386, 0.0245428, 0.00710099, 0.000112951, 0.000117222, 0.0293721, 0.00016956, 0.0180899, 4.12346e-06, 0.00078692, 0.000158923, 0.0134079, 0.0305984, 4.49839e-05, 0.00263064, 0.00292569, 0.00203238, 0.017169, 0.000220931, 0.123046, 0.000207955, 7.43336e-05, 0.0172892, 0.00139245, 0.223386, 0.000183165, 0.000418319, 0.00102132, 0.000576292, 6.94214e-06, 0.0469607, 2.84531e-05, 0.000593348, 0.00276771, 0.00069356, 0.00178553, 6.58826e-06, 0.00308764, 0.000531875, 1.83493e-07, 5.96315e-06, 0.00167116, 0.00084332, 0.000194465, 6.43939e-05, 0.00201767, 0.00163908, 3.19069e-05, 0.0226464, 0.00115657, 0.000850113, 0.000533881, 2.29994e-05, 3.62086e-05, 0.000366299, 0.10905, 2.31936e-05, 0.000504016, 0.00511315, 0.040291, 8.2822e-05, 0.0265812, 0.00251949, 0.633334, 0.00141468, 1.26084e-05, 8.29132e-05, 0.000101538, 0.00137941, 4.03015e-05, 0.00310858, 0.00616134, 0.00519421, 0.00150005, 0.000136335, 0.00016128, 7.11051e-05, 0.000218231, 0.00035392, 0.00148934, 6.07769e-05, 0.000283501, 1.81238e-05, 0.00683618, 0.0153657, 9.69282e-05, 0.00103645, 0.00196901, 0.00130667, 0.000617759, 0.000589285, 0.00425305, 0.00108045, 5.99169e-05, 0.00120762, 0.000747274, 1.03621e-05, 0.00491635, 0.000213026, 0.067325, 0.00202005, 1.74527e-05, 0.00963645, 0.00766829, 0.0021894, 0.00198882, 1.21556e-06, 0.000381541, 2.59649e-05, 0.00414563, 0.000937585, 0.00376576, 9.8046e-05, 0.00184014, 0.00727371, 0.000227209, 0.0147809, 0.000151495, 0.000511387, 0.0820332, 0.0136848, 0.000582104, 3.16842e-05, 0.000195939, 0.0391684, 0.000551577, 0.074382, 0.000954884, 0.0281147, 0.000261419, 0.000154929, 0.000119419, 0.0201258, 0.0573592, 0.000286688, 0.000278933, 0.00128995, 5.96899e-05, 0.000111202, 0.109798, 2.41892e-05, 0.00376636, 0.0139996, 0.00257616, 3.38323e-05, 0.000352247, 0.000332327, 0.00027496, 4.38117e-05, 0.000322345, 0.0213569, 2.24847e-05, 0.0367522, 0.000150906, 3.79968e-05, 0.00866311, 0.0133283, 2.54953e-05, 0.0012242, 0.000488975, 0.00835136, 5.6915e-05, 1.39304e-05, 0.000123466, 0.0933792, 0.00912074, 0.000105051, 3.10912e-05, 0.312335, 0.00493653, 0.00408567, 0.000123634, 0.00648313, 0.0060388, 0.0151249, 0.000249494, 0.00881769, 0.0443835, 0.000302048, 0.0148538, 0.00059659, 0.0276766, 1.6993e-05, 0.710075, 0.000335147, 7.4766e-05, 1.63472e-05, 0.00164538, 0.000212225, 1.7083e-05, 0.000152778, 0.000125165, 0.001976, 5.08364e-05, 0.007717, 7.09645e-05, 0.0334573, 5.09435e-05, 0.00123988, 1.52463e-05, 0.0321567, 5.48258e-05, 0.000355483, 0.00121258, 2.49229e-05, 0.00318118, 0.00137222, 0.00162397, 0.000937861, 3.11791e-05, 0.000171446, 0.000189181, 0.00168137, 0.00123549, 0.00210851, 0.000254292, 0.000420091, 0.00109599, 0.000327945, 0.0407798, 1.27419e-05, 6.51907e-05, 0.0111275, 0.000242079, 9.24735e-06, 0.011485, 6.89149e-05, 5.50909e-05, 0.00068756, 1.37815e-05, 0.00020156, 0.000249486, 0.00405363, 0.000185963, 0.000427773, 0.000462914, 1.87303e-07, 3.31144e-05, 5.23371e-07, 1.37587e-05, 2.41396e-06, 1.69499e-06, 5.15217e-06, 0.000460911, 1.06383e-06, 1.04087e-05, 2.43939e-06, 3.11792e-06, 4.07773e-05, 0.000106851, 2.11539e-07, 7.04064e-07, 6.04858e-05, 8.91215e-06, 2.88211e-06, 4.87253e-07, 0.000899104, 2.12469e-05, 0.960883, 2.75458e-07, 2.66301e-05, 2.84961e-05, 1.77408e-05, 1.6432e-05, 3.98505e-05, 0.000403693, 3.07816e-05, 0.027278, 7.45977e-08, 0.000171766, 1.87415e-06, 3.93902e-05, 0.000126695, 0.00764235, 3.65531e-08, 0.000108071, 0.000340822, 1.03099e-05, 1.236e-06, 1.25154e-06, 1.62174e-05, 2.96779e-05, 1.77054e-07, 1.96063e-07, 3.76494e-06, 4.9137e-06, 2.06857e-05, 1.81563e-05, 0.000110109, 1.35185e-05, 6.3757e-06, 6.20252e-06, 2.21654e-05, 7.11291e-06, 0.000107109, 3.53048e-07, 0.000107517, 9.8599e-07, 0.000216255, 1.02083e-05, 0.000191973, 1.19248e-06, 0.000394728, 0.000343793, 2.73654e-05, 0.000716608, 3.58214e-07, 2.0473e-06, 0.000191015, 0.00611411, 2.50127e-07, 0.00203114, 0.000301405, 0.000165523, 0.000334263, 2.38782e-05, 0.000460268, 5.15675e-05, 0.000552039, 0.0101524, 6.94102e-07, 0.000198129, 2.83347e-06, 0.000795017, 0.0122051, 3.63419e-06, 0.000290447, 0.000157087, 5.19854e-06, 0.00108856, 3.26339e-08, 2.06163e-05, 4.37503e-06, 0.000236836, 1.01836e-05, 0.000372176, 2.56547e-05, 0.0221835, 1.69105e-05, 3.33914e-05, 4.44601e-05, 0.00131123, 2.32862e-05, 5.57634e-05, 0.000158584, 8.5176e-06, 9.26429e-05, 0.000591125, 0.000263717, 1.25154e-06, 0.00119017, 0.000678137, 0.000446496, 2.57072e-07, 0.000125783, 1.90438e-06, 6.85187e-06, 0.000549256, 7.95432e-05, 0.00747503, 0.926844, 5.33467e-05, 0.000282226, 0.00239624, 0.16913, 0.00366106, 0.000555541, 6.22286e-05, 0.000821962, 0.0200465, 1.90485e-05, 0.00193741, 0.0245157, 0.00120272, 5.83544e-05, 0.0066611, 0.00547818, 0.000398698, 2.61857e-07, 0.0208841, 0.000324911, 2.69985e-05, 4.75462e-05, 0.00583974, 0.00333466, 0.000127601, 0.000391615, 2.9782e-05, 6.02551e-05, 0.0347449, 0.0013559, 0.00244366, 0.00245175, 0.00169861, 0.000215885, 3.88141e-06, 0.000146797, 0.00833083, 0.0661406, 0.0801496, 0.000185113, 0.0329386, 3.45846e-05, 3.74859e-05, 0.00307437, 0.000479584, 0.000409071, 8.6429e-05, 0.018033, 7.03646e-05, 9.52391e-06, 1.1442e-05, 0.453907, 0.000115511, 5.27709e-05, 0.00243411, 3.74606e-05, 0.0035258, 8.10122e-06, 5.63744e-05, 4.16015e-05, 0.00021681, 0.0112295, 0.00292599, 0.00422214, 4.46605e-05, 0.000118514, 0.000784396, 0.000658417, 6.58046e-05, 0.00317173, 0.000113274, 0.01285, 1.96802e-06, 0.00189987, 0.00874199, 0.00181971, 0.000207146, 7.70333e-05, 0.000436817, 4.31241e-05, 0.716222, 0.000905786, 0.000249811, 0.00821159, 9.84286e-05, 0.00253578, 0.000479148, 0.000331113, 1.60484e-06, 0.00282729, 1.8331e-05, 4.81206e-05, 0.000657298, 0.000839732, 0.0307182, 0.000973946, 0.00226545, 0.00696592, 1.44221e-06, 0.000304523, 0.0295347, 0.00234378, 0.0936875, 0.000200097, 1.14535e-05, 2.25442e-05, 0.000189406, 5.07032e-05, 0.000589688, 0.00105097, 0.00219676, 0.00279262, 9.73582e-05, 0.0164431, 0.00351529, 0.00208491, 4.86261e-05, 9.84532e-06, 0.000539089, 0.00267792, 0.00842115, 0.00413775, 0.00217523, 0.00163346, 0.000858024, 0.00315549, 0.0124298, 0.000142852, 0.00240117, 0.00103164, 8.62605e-05, 0.000379209, 0.000347443, 4.49872e-05, 0.000368599, 0.000288879, 0.0021179, 3.68653e-06, 8.69825e-05, 7.45124e-05, 5.35482e-06, 2.47377e-05, 0.000444259, 4.73985e-05, 0.900217, 3.8801e-05, 0.000329459, 7.29902e-06, 3.7039e-06, 0.00224328, 0.00322497, 0.00063843, 1.56523e-05, 2.13544e-05, 1.01867e-05, 0.000364608, 1.90732e-06, 0.00211248, 4.43922e-06, 0.000152961, 0.00120167, 0.00143112, 8.01921e-06, 0.00047465, 0.00970861, 0.00016028, 0.000149195, 1.81241e-05, 0.00059492, 6.28811e-05, 9.10615e-05, 0.000489821, 1.48951e-05, 7.58268e-05, 3.90343e-05, 0.0636397, 5.75078e-05, 1.5125e-05, 3.96813e-06, 8.30622e-05, 9.22125e-06, 1.61137e-06, 0.000379977, 0.000185057, 3.99829e-05, 0.000382608, 0.000569389, 4.46873e-05, 0.00601576, 1.80704e-05, 3.38251e-06, 0.000297699, 2.49273e-06, 2.34389e-05, 8.96723e-05, 5.30497e-05, 0.00729375, 8.79986e-06, 0.000626107, 0.00190138, 0.00624868, 5.74496e-05, 0.000127143, 7.10914e-07, 0.00223327, 0.00186455, 0.911844, 0.0280662, 0.000839562, 8.22344e-05, 1.97394e-06, 0.00040967, 2.15311e-05, 3.42471e-05, 7.67371e-05, 0.00103658, 0.001352, 4.0259e-05, 0.000743141, 0.00364524, 4.95056e-05, 1.85409e-05, 0.000746018, 7.38349e-06, 4.39968e-05, 0.00117578, 0.000895574, 0.000341397, 1.63615e-05, 0.000414441, 1.18566e-05, 4.9823e-06, 8.21261e-05, 4.39658e-07, 0.000176708, 4.98424e-05, 0.000715883, 1.52478e-06, 5.45687e-06, 2.40077e-06, 0.00175838, 0.000139896, 0.000183166, 8.57927e-06, 6.46772e-07, 1.30371e-05, 8.84676e-05, 4.95152e-05, 0.000450073, 0.0164699, 0.000210617, 2.40701e-05, 3.53283e-05, 4.84123e-05, 0.00290064, 0.00101769, 0.000124663, 0.00301874, 2.85746e-06, 2.82874e-06, 6.61084e-05, 4.72847e-05, 0.000429898, 4.92679e-05, 2.28277e-06, 4.23815e-06, 0.000299787, 2.36378e-06, 1.2859e-05, 1.92258e-05, 4.62881e-05, 0.000300544, 2.78808e-06, 2.14605e-05, 0.00502293, 0.000318405, 0.000134688, 0.000179439, 4.0844e-06, 5.8571e-05, 4.39902e-05, 8.43654e-06, 2.00831e-05, 4.57546e-05, 0.000485299, 4.31111e-05, 1.24607e-05, 3.53776e-05, 1.23501e-05, 1.24581e-06, 8.07402e-06, 8.16516e-06, 6.26264e-06, 4.67781e-08, 2.21237e-05, 5.62087e-07, 0.00253227, 2.22899e-07, 2.19952e-05, 4.15495e-06, 4.43227e-06, 0.000101796, 1.47817e-05, 4.03126e-05, 1.83463e-05, 0.000484774, 1.90168e-05, 5.68202e-06, 2.9716e-06, 0.000131621, 0.000234169, 0.000138988, 0.000374801, 4.26675e-06, 8.73954e-07, 0.0002446, 0.000430339, 1.70498e-06, 3.65632e-06, 1.13644e-05, 0.987303, 8.81621e-05, 0.000429137, 2.94455e-05, 0.85404, 0.00102732, 0.000353811, 0.000629049, 0.00164571, 0.00432708, 0.000322842, 7.78785e-05, 3.02458e-05, 0.000375771, 0.00553467, 0.000347797, 6.25297e-05, 0.00173365, 0.000103935, 0.000476735, 0.00768624, 4.6448e-06, 2.93519e-05, 7.53928e-05, 0.000159902, 3.81013e-05, 0.00294544, 0.00129934, 0.000121318, 7.7421e-05, 3.9432e-06, 0.000779413, 0.00237856, 1.03151e-05, 0.00673123, 1.22947e-05, 0.000122518, 0.000171755, 0.000121174, 1.86879e-05, 0.00620762, 6.07784e-05, 0.00353672, 0.0765273, 0.00220243, 1.15758e-05, 0.000212994, 0.00120673, 4.29309e-05, 0.000177674, 0.00664306, 0.000478648, 0.000143262, 0.00147426, 0.000135916, 0.000162114, 0.00203429, 0.000534597, 0.000811432, 0.000415027, 4.67413e-05, 0.000183834, 0.00023661, 2.33756e-05, 1.01623e-05, 0.00214457, 0.00040186, 0.00105837, 4.38379e-05, 4.4035e-05, 0.000127841, 0.000902461, 5.84014e-05, 0.0137974, 4.94618e-05, 0.00634612, 6.95644e-05, 0.86419, 0.00289061, 8.60206e-05, 0.00116974, 0.000773566, 1.41516e-05, 8.58911e-05, 0.0231014, 8.90855e-06, 0.000714118, 0.000186736, 0.000877173, 3.54151e-05, 0.00943081, 2.71024e-06, 0.0109015, 0.00407662, 8.91125e-06, 0.000242833, 0.00333538, 1.36708e-05, 0.000322224, 8.10862e-06, 0.0106816, 1.45518e-06, 7.6912e-06, 0.00254539, 0.000104255, 0.00231797, 2.6628e-05, 1.12786e-05, 2.70028e-06, 0.00234318, 0.000804449, 0.00125266, 0.000400489, 9.38665e-05, 3.85096e-06, 8.41146e-05, 0.0021653, 0.00108366, 0.000319942, 0.00091975, 4.08588e-05, 0.00300251, 0.000258126, 3.068e-05, 3.94615e-07, 6.91138e-05, 0.00226669, 0.00105301, 1.1816e-05, 0.0227204, 1.00372e-06, 0.000690658, 6.79657e-05, 0.00156317, 6.5294e-07, 0.019864, 6.55073e-07, 9.71677e-06, 2.48639e-05, 2.89201e-05, 0.000143838, 6.25836e-05, 9.33023e-06, 3.64007e-06, 2.02009e-06, 0.000531067, 0.000567802, 4.21978e-05, 0.000564553, 1.5312e-05, 5.58189e-05, 0.0108437, 4.45389e-06, 4.39566e-05, 0.0021238, 0.000122196, 4.46943e-06, 9.39855e-05, 4.49352e-06, 0.0101324, 0.0003163, 9.56112e-07, 0.000134794, 5.08243e-05, 0.0012151, 6.3432e-05, 0.000148798, 1.62182e-05, 0.00510486, 3.87345e-05, 3.61156e-07, 5.27255e-06, 0.00105663, 0.924897, 0.000912537, 1.30823e-05, 2.22205e-05, 5.87463e-05, 0.000394205, 8.67987e-07, 0.000687406, 4.11963e-05, 3.88136e-05, 8.47525e-05, 4.41435e-05, 6.45206e-05, 9.72723e-05, 0.0158785, 0.000409092, 0.000298649, 5.71165e-06, 9.89439e-06, 4.05908e-06, 0.000261873, 0.00159044, 0.000127516, 0.031494, 0.0314191, 0.00202146, 0.00192303, 0.0182093, 0.000754009, 0.0329904, 0.539312, 0.00011743, 0.0179833, 0.000157735, 0.00037711, 0.000258713, 2.22246e-05, 4.36575e-05, 0.00393777, 0.00257851, 4.35908e-05, 0.000148693, 0.0573748, 3.79556e-05, 0.00730448, 0.00289943, 0.00385297, 0.0056806, 0.000123868, 0.016346, 0.00128758, 3.12499e-05, 0.000198956, 2.77188e-05, 3.11142e-05, 7.41944e-05, 0.000982676, 0.000252415, 0.00464785, 0.000173071, 1.21665e-05, 0.000499775, 0.00102407, 0.00298303, 0.0999068, 0.00440586, 4.61109e-05, 0.0074805, 0.000811496, 0.00138468, 0.065239, 0.00370491, 0.000111921, 0.00500768, 0.000177251, 0.000174751, 0.00897356, 1.64966e-06, 4.00663e-05, 4.26004e-05, 0.00538819, 0.00089155, 1.98212e-05, 3.8817e-05, 0.00479633, 0.000726413, 0.00517782, 0.000184576, 0.000138115, 0.000138049, 0.0038683, 0.000828995, 0.127322, 8.47342e-05, 0.00101261, 0.000139882, 0.0121463, 3.04805e-05, 8.53097e-05, 1.29643e-05, 0.0733171, 0.0105553, 0.00423729, 0.00648643, 6.44155e-06, 8.23094e-05, 0.407672, 0.00447902, 0.000601691, 0.000401925, 0.000396175, 0.009328, 1.29655e-05, 9.47796e-07, 3.82517e-05, 0.000927429, 0.000242166, 0.000205103, 1.44197e-06, 0.00149505, 0.1228, 0.0359184, 0.0233735, 0.000160162, 0.00990549, 0.000789739, 0.0150316, 4.34187e-05, 0.000192486, 0.00631615, 0.00905066, 0.00313928, 0.00291798, 2.69386e-05, 0.00634979, 0.00151166, 5.24448e-05, 0.0689153, 0.000122014, 0.00111086, 0.000463391, 0.00331782, 0.000105677, 0.00123212, 0.00610067, 0.00599769, 0.000378889, 0.000307754, 0.00198361, 0.033403, 0.094569, 9.59557e-05, 9.18797e-06, 0.000370446, 0.00166521, 0.0139587, 0.00992985, 0.0193808, 0.000442261, 0.000168851, 0.000755342, 3.48162e-05, 0.000152497, 0.000406833, 0.0126832, 0.000238296, 0.000306682, 0.00332674, 0.000893981, 0.025855, 0.00196351, 0.000347468, 0.0104901, 5.74978e-05, 0.00025415, 0.00243229, 4.82324e-05, 0.00289439, 0.00237848, 0.00931408, 0.0074032, 0.015734, 9.73412e-06, 4.80434e-05, 0.00167448, 0.0635554, 2.48705e-06, 1.85689e-05, 0.00689109, 0.0111869, 3.10351e-05, 0.00164966, 5.46416e-06, 0.00341474, 0.000169618, 0.00175692, 0.00224225, 3.52626e-06, 0.00114524, 0.00391467, 0.11215, 1.06941e-05, 2.5565e-05, 0.000109809, 0.00560542, 0.000542117, 0.00470967, 0.323866, 1.02983e-05, 0.000300602, 0.132252, 0.0505506, 0.000183025, 3.08849e-05, 0.00302125, 0.00141351, 0.043944, 0.00896993, 0.00275497, 0.0105176, 0.0219871, 0.00321966, 0.000380122, 0.000327038, 0.0478673, 0.0138036, 0.00913533, 8.85604e-06, 0.00193606, 0.00112436, 0.0188074, 0.255254, 0.00016376, 7.48709e-05, 0.000868831, 0.0543695, 0.000179394, 0.018722, 0.00105465, 0.00559472, 0.000555503, 0.000357946, 0.0582159, 0.000601961, 0.0501166, 0.0171797, 0.000908199, 0.0114616, 0.000152955, 0.00428875, 0.000169914, 0.000103695, 0.00511389, 0.0020082, 0.000165009, 0.000530205, 0.0195669, 0.0308734, 0.0010241, 0.000633291, 0.000218855, 0.0103766, 0.000720982, 0.0237633, 0.00887525, 6.74827e-05, 0.000196819, 0.000630228, 0.0038318, 0.0226822, 1.54444e-05, 0.00404619, 0.0789054, 0.105775, 0.000500122, 0.00470862, 0.00512736, 0.0012488, 0.00262702, 0.000787695, 0.00395985, 0.000238098, 0.340815, 0.00458061, 4.12822e-05, 0.000268, 3.72777e-06, 0.0282613, 0.000296478, 6.35393e-05, 9.80279e-06, 0.012017, 0.00151411, 0.000227564, 0.00100013, 0.00031217, 0.00010466, 0.371554, 0.00227485, 0.000484757, 9.45801e-06, 4.77369e-05, 0.0114878, 0.000122714, 4.81217e-05, 0.0135532, 0.000196776, 0.00360726, 2.61237e-05, 0.000340998, 0.00127349, 0.000109689, 0.00145293, 0.000519115, 7.60422e-06, 0.000101856, 1.56287e-05, 1.61842e-05, 0.000345459, 5.09812e-06, 0.151465, 0.00561902, 5.52466e-05, 0.000150794, 1.30224e-07, 1.87092e-06, 0.00295726, 0.00240554, 3.73827e-06, 2.75565e-05, 1.95769e-05, 0.000158207, 3.44886e-05, 2.23985e-05, 0.030027, 9.05991e-05, 0.000398529, 0.000343926, 0.000114813, 4.6068e-05, 8.06099e-05, 4.31696e-06, 5.67565e-05, 0.000131289, 6.64384e-07, 7.51317e-05, 4.56596e-05, 0.000643042, 0.00353389, 0.000161054, 1.59944e-06, 0.000150204, 0.0018035, 1.44856e-05, 4.42082e-06, 0.000616464, 9.07485e-07, 1.276e-05, 0.000229805, 6.67744e-07, 1.31625e-06, 4.0701e-05, 9.11933e-05, 0.000398379, 5.09894e-05, 2.88746e-05, 2.47778e-06, 0.000781762, 0.00307452, 7.7328e-07, 1.48324e-05, 0.00321194, 0.000111986, 0.000329615, 6.00636e-05, 8.15107e-05, 0.000101644, 0.00031296, 0.000341557, 1.56422e-05, 5.12144e-06, 0.000313823, 1.28723e-05, 0.000114394, 4.96307e-06, 0.000202445, 0.000900563, 0.0010076, 0.950629, 0.000672463, 0.00037074, 0.00322852, 0.000489998, 0.00138401, 0.00184368, 0.00172527, 0.00232951, 8.60045e-05, 0.000103409, 4.06574e-05, 7.18382e-06, 0.00371164, 1.38328e-05, 0.0130328, 0.00122964, 1.27035e-05, 0.0291633, 0.00458457, 0.00139699, 0.00339122, 0.00469614, 3.45075e-06, 7.37784e-05, 0.00375217, 0.00511432, 8.3081e-05, 0.0689141, 0.000341116, 0.000234531, 0.158131, 0.000222956, 6.86339e-05, 0.389799, 6.99764e-06, 0.0041345, 6.25963e-05, 1.57697e-05, 0.000187285, 0.00076429, 0.0434235, 0.0156617, 0.000398015, 0.00143705, 0.00379662, 0.000955689, 0.000917555, 0.00201681, 0.000769939, 0.00507487, 0.0007072, 8.48702e-05, 0.000328958, 0.0014615, 0.0002973, 0.00937292, 9.01885e-05, 1.81804e-05, 6.29549e-05, 0.00515181, 0.0075939, 0.000233867, 0.147747, 4.42918e-05, 0.000191441, 0.000598894, 0.0120973, 0.00100711, 0.00355096, 0.00125667, 0.00832444, 0.00977281, 7.86796e-05, 0.000437234, 0.00568722, 0.00728534, 0.0233009, 0.00045579, 0.00108893, 0.00206397, 5.86502e-06, 3.56316e-07, 3.41327e-05, 0.000392354, 2.14202e-05, 1.02814e-05, 1.71957e-05, 1.75407e-05, 1.18235e-06, 5.422e-05, 0.000108071, 6.54377e-05, 5.8327e-05, 2.49765e-07, 4.29452e-05, 2.40542e-07, 1.96407e-05, 0.00505796, 0.000154708, 9.71746e-05, 0.000186406, 4.34553e-05, 1.49686e-05, 2.26156e-06, 1.33576e-05, 0.00288, 4.34793e-05, 4.12368e-06, 1.05605e-06, 3.88238e-06, 6.7219e-05, 2.61861e-08, 2.17662e-06, 9.74076e-06, 0.00516068, 0.00011608, 3.22985e-05, 7.89115e-06, 4.6519e-05, 1.15431e-05, 0.000829485, 6.93879e-06, 5.21308e-07, 6.08418e-07, 2.53557e-05, 8.36144e-05, 2.59315e-07, 5.94136e-06, 0.980216, 0.00364994, 7.64249e-06, 6.63828e-05, 3.82046e-06, 6.97045e-05, 3.46959e-06, 9.07199e-07, 8.5024e-05, 9.40581e-06, 7.61765e-06, 1.43335e-06, 2.94773e-05, 3.5594e-06, 1.20896e-05, 7.42959e-05, 0.000551536, 0.000179463, 0.00528553, 0.00033407, 1.00558e-06, 0.000111002, 0.000174404, 0.152246, 1.49212e-05, 8.19497e-06, 0.000110353, 0.0678637, 4.50698e-05, 0.000296619, 2.24829e-05, 0.00218984, 6.68495e-05, 0.000577276, 0.00893942, 8.02683e-06, 0.000352924, 0.000418488, 0.410534, 0.0101737, 0.000783056, 0.000110296, 0.00247655, 0.0156463, 0.00729219, 0.00017294, 0.0016027, 0.00197942, 0.00248786, 0.0117393, 0.251667, 0.00293213, 2.66822e-05, 0.000217641, 0.00350704, 0.000254866, 0.0141366, 0.000472334, 9.99399e-06, 0.000576862, 0.000150828, 3.68321e-05, 0.00068376, 0.00282028, 4.6627e-05, 3.22705e-05, 0.000567164, 5.85279e-05, 0.000175397, 0.010975, 1.73056e-05, 0.00111812, 0.000994886, 0.000513829, 0.000123878, 1.90761e-05, 0.00190325, 0.000213058, 0.000909484, 4.44691e-05, 0.000221819, 1.28653e-05, 1.88079e-06, 0.000221713, 2.73183e-05, 0.987719, 1.16316e-06, 0.000204505, 3.48615e-06, 6.4668e-06, 2.96872e-05, 5.02437e-06, 2.96182e-06, 4.55241e-05, 3.30765e-05, 3.90472e-07, 4.63102e-07, 1.11558e-05, 4.29713e-05, 0.000147436, 0.000171367, 0.000103933, 8.52325e-07, 3.77151e-05, 5.98757e-06, 0.0017335, 6.37394e-05, 0.000257579, 1.01819e-05, 3.5406e-06, 1.28806e-06, 0.000125741, 2.60778e-05, 9.58307e-08, 0.000169812, 2.09041e-06, 1.41587e-05, 0.000141879, 1.07191e-06, 0.000307991, 1.35309e-05, 6.04491e-06, 1.47142e-05, 7.9816e-05, 1.35051e-07, 2.69764e-06, 2.54126e-05, 0.000377846, 1.23109e-05, 4.07709e-06, 1.16981e-05, 0.000225682, 3.39496e-07, 0.0057026, 1.21412e-06, 3.3614e-05, 8.53665e-06, 2.29428e-06, 5.35975e-08, 8.74855e-05, 0.00096755, 8.43607e-05, 0.000146923, 0.000273983, 1.30934e-05, 1.99938e-05, 1.38312e-05, 0.000870432, 0.00106592, 4.31595e-06, 0.0013832, 0.000879489, 5.6233e-05, 0.000136366, 1.53341e-07, 4.41621e-07, 8.87484e-05, 0.000173022, 8.25478e-05, 3.78476e-07, 0.000109477, 0.000211975, 0.000126409, 1.84011e-05, 0.000565508, 0.000320857, 0.00104747, 4.6954e-06, 1.4534e-05, 3.28798e-06, 7.73158e-05, 0.0011412, 0.759321, 0.0204672, 0.000537298, 0.00323739, 0.000571466, 3.25825e-05, 8.36364e-06, 0.0015478, 6.04023e-05, 0.000184216, 5.53283e-06, 1.66864e-06, 6.31028e-06, 0.00347747, 0.000883697, 0.000371823, 0.0540246, 0.00854085, 0.00394921, 3.41245e-06, 4.09956e-07, 0.00146326, 0.000161649, 0.000261311, 0.000581719, 0.000109364, 0.000423012, 4.63978e-06, 9.59955e-06, 0.000188438, 0.000310496, 0.130352, 0.00012367, 1.46646e-05, 4.89151e-07, 0.000333933, 0.00323496, 0.00748899, 0.015951, 0.00013061, 0.000164172, 2.39005e-05, 3.54864e-05, 7.44106e-05, 0.00101412, 0.000713124, 1.60138e-05, 0.00114383, 0.000943696, 7.83051e-05, 0.000274759, 0.00202453, 0.000773422, 0.0125991, 6.63727e-05, 0.00498811, 1.05747e-05, 7.65362e-05, 0.0052302, 0.000107691, 0.00156034, 8.42779e-05, 0.00302817, 0.0264369, 0.0009573, 9.97913e-05, 0.000829368, 1.24987e-05, 3.13941e-05, 0.000918127, 0.000369182, 2.85529e-05, 0.000428047, 0.152101, 0.000669032, 0.00502334, 0.00125686, 0.000788526, 8.48772e-05, 0.000158385, 0.000808015, 0.00535268, 0.00370566, 0.619811, 0.0020807, 0.000602897, 0.00238879, 8.9951e-05, 0.00801975, 0.0866519, 0.0108544, 0.00295008, 0.000204652, 0.000338636, 0.00145994, 1.75809e-05, 1.15862e-05, 0.00212107, 0.000483646, 1.69554e-05, 7.06371e-05, 2.93114e-05, 6.6221e-06, 1.49463e-06, 0.00215276, 1.52825e-05, 0.000117298, 1.1e-05, 0.000392924, 4.83054e-05, 0.00053803, 1.73669e-06, 7.5691e-06, 1.77578e-05, 1.51772e-06, 1.17091e-06, 1.64341e-05, 3.2224e-06, 0.000125903, 0.000142242, 7.30791e-07, 7.4454e-06, 0.000199895, 3.48927e-06, 2.54335e-06, 4.32769e-05, 2.21488e-06, 4.79744e-05, 0.0254309, 3.22579e-05, 8.28033e-05, 0.000124679, 2.47937e-06, 2.63551e-05, 7.24203e-07, 1.43096e-05, 0.000746061, 0.00481296, 8.20438e-06, 5.98771e-07, 8.98461e-06, 0.000156274, 2.77788e-06, 0.956593, 0.000168099, 1.5599e-06, 4.33813e-07, 8.24235e-06, 0.000494895, 2.63538e-06, 0.00120652, 0.0011484, 3.22767e-06, 1.68661e-06, 4.49111e-07, 2.23826e-07, 0.00421787, 5.89979e-05, 0.000539467, 5.64046e-05, 1.36529e-06, 2.28501e-05, 2.18029e-06, 1.22656e-05, 0.00126959, 0.000918662, 0.00793484, 4.4886e-05, 5.49134e-05, 0.000798614, 9.80251e-05, 7.86549e-05, 0.0078591, 0.000821371, 3.52096e-07, 0.00135158, 0.00134718, 7.84287e-05, 0.00347373, 0.000124678, 0.00331548, 0.0118817, 9.47152e-05, 8.16131e-06, 4.6269e-05, 0.929685, 0.000449564, 8.89002e-06, 0.0013001, 0.000422626, 2.89495e-05, 0.00015841, 1.41101e-05, 0.00399092, 0.00285052, 0.000637837, 1.39951e-05, 0.00172396, 9.14221e-06, 0.000120447, 0.00289172, 0.00228382, 4.26141e-05, 2.52891e-05, 5.31715e-05, 0.00239949, 0.000198459, 0.000202748, 6.31179e-05, 0.000381464, 0.00333772, 4.19278e-06, 0.000185018, 2.17014e-05, 8.9833e-06, 0.00042338, 0.000274916, 0.000183511, 0.000187147, 0.000303256, 8.49791e-05, 0.00108922, 0.00165908, 0.00026755, 2.81661e-05, 3.91496e-06, 4.76602e-05, 0.000332522, 0.000225776, 0.000213632, 0.00064789, 1.9785e-05, 1.27926e-05, 0.000242899, 0.000519132, 0.00180709, 0.000606684, 0.015078, 1.88921e-05, 0.000499366, 0.00505859, 1.62624e-05, 0.00899928, 0.000113715, 0.00104129, 5.16046e-05, 0.0494343, 6.24444e-06, 1.73912e-05, 0.000778818, 0.00284035, 0.000812244, 7.04733e-05, 0.0174798, 0.000257386, 9.94766e-05, 0.000451436, 0.00117564, 0.0654707, 2.34676e-06, 0.000181725, 3.39697e-05, 0.00131877, 3.48732e-05, 3.80591e-06, 0.000634624, 0.00496871, 0.00360456, 0.801039, 6.80148e-06, 9.15689e-05, 0.000364993, 5.02366e-06, 0.0002693, 4.22519e-05, 0.000176599, 0.00165365, 0.000515916, 0.00117229, 0.000607643, 0.000196585, 0.000141104, 0.00129101, 0.000313933, 0.000273662, 1.41052e-05, 0.00179944, 0.000126835, 0.000650457, 0.0033958, 2.4944e-05, 0.000976754, 0.0124847, 0.0889412, 0.000287495, 0.000275128, 4.15409e-05, 0.00847689, 0.000706653, 0.000132277, 0.000139421, 0.0017816, 0.000955939, 0.000274882, 0.00146839, 0.000589945, 0.237942, 0.00125089, 0.00010479, 0.00269618, 0.000569142, 0.000535077, 0.00107987, 0.000568907, 0.00012988, 0.0038759, 0.000542725, 0.00127305, 7.87723e-05, 0.000182838, 0.00244898, 0.00247886, 0.227766, 0.0710133, 0.000232961, 0.000114483, 0.00365246, 0.188955, 0.00262869, 2.90533e-05, 5.95087e-05, 0.00306103, 0.00126976, 0.00596232, 2.7084e-05, 0.000252908, 0.000109971, 0.000898649, 0.00122538, 0.000374626, 0.00124336, 7.88324e-05, 0.0332834, 0.000177044, 0.0411651, 0.0143063, 0.00189394, 4.80069e-05, 0.00447686, 0.000386166, 0.000207919, 0.0162557, 0.000963902, 0.00294336, 0.000451154, 0.0021722, 0.158712, 0.00528903, 0.0219807, 0.000219388, 0.000558334, 0.000550728, 0.160331, 0.000437625, 0.000502855, 8.237e-05, 0.000159942, 0.00440698, 0.00251776, 0.000249967, 0.000286408, 0.00450498, 0.00802651, 0.000256257, 0.000742709, 7.32709e-05, 0.0333007, 0.000400981, 0.00283973, 0.000579002, 1.57269e-05, 0.0133198, 0.0487912, 0.0326244, 0.00858963, 0.000146478, 1.91681e-05, 0.124022, 0.00246038, 0.0268233, 0.00222943, 0.00184737, 2.56893e-05, 0.00138688, 0.000951848, 0.0011091, 0.00546575, 0.00633182, 0.000337227, 0.000185501, 0.00292652, 3.71637e-05, 0.00283592, 0.000854371, 0.0240517, 0.000124184, 0.000410151, 0.00255549, 0.000596397, 0.000256948, 0.00793872, 0.17319, 0.00461879, 0.000118633, 0.0710592, 0.000544056, 0.000652606, 0.000426687, 0.0202969, 0.00281363, 0.00593937, 0.000404239, 0.00246433, 8.30009e-07, 4.36686e-05, 3.99098e-05, 0.00050359, 5.63552e-06, 7.34384e-05, 0.00051771, 0.000311964, 0.00445179, 0.00134203, 8.73793e-05, 0.000247356, 0.00041109, 5.10104e-05, 0.0136675, 0.0121955, 0.000166304, 0.0108291, 0.0052992, 0.000175425, 0.000332845, 0.389711, 0.00437276, 0.00769929, 0.000116276, 0.00321803, 4.081e-05, 0.00126027, 0.000568228, 0.000208154, 8.0133e-05, 3.59278e-05, 0.342036, 0.00696144, 0.0366896, 0.000170747, 0.000196377, 5.4992e-05, 0.00345135, 0.0172704, 0.00123913, 0.000132208, 3.89683e-05, 3.7082e-06, 1.33235e-05, 0.00269294, 1.85952e-05, 0.000463462, 0.00101639, 1.78215e-05, 8.56869e-05, 0.000406752, 0.000252077, 0.0199581, 0.0955011, 1.4186e-05, 0.000485586, 0.000197744, 0.003748, 1.00076e-05, 1.45761e-06};
+
 #endif // CONFIG_CONSTANTS_H
